refactor(steerer): extract _handle_drift + dispatch + ladder into DriftObserver (bucket 3c of 3 — final steerer split)
Final cut of the Wave-C steerer split. Bucket 3c moves the dispatch +
intervention ladder + promotion + cancel + refine-outcome bookkeeping
machinery onto DriftObserver. DefaultSteerer becomes the thin router
that owns shared mutable state (sinks, planner, adapter, control_channel,
ContextVar for active session, plan locks, observation_only flag,
REFINE_FAILURE_THRESHOLD / PROGRESS_STALL_THRESHOLD_SECONDS class
constants) plus the shim layer that tests + the planner +
goldfive_planner hook + executors historically hooked at the
steerer.X bare-attribute names.

Methods moved (verbatim, with self.X -> self._steerer.X for router
state; method bodies otherwise byte-equivalent):

* _handle_drift (-> DriftObserver.handle_drift, 750-line central
  drift-routing method)
* The intervention ladder: _LADDER / _LADDER_BY_VALUE /
  _ladder_level_for / InterventionLevel ladder dispatch
  (_dispatch_nudge, _dispatch_goldfive_steer_control,
  _dispatch_goldfive_pause_control, _dispatch_pause_escalate)
* Adapter cancel tagging (_tag_adapter_cancel_reason,
  _tag_adapter_cancel_reason_for_promotion,
  _write_adapter_cancel_reason, _request_adapter_cancel,
  _ADAPTER_CANCEL_REASON_USER_STEER)
* Cooperative cancellation (_is_late_drift_for_terminated_invocation,
  _resolve_active_invocation_ids, request_invocation_cancel,
  _should_request_cancel_for_drift, _cancel_reason_for_drift,
  _cancel_inflight_for_revision)
* Promotion policy + handler (_GOLDFIVE_STEER_ELIGIBLE_KINDS,
  _severity_meets_promotion_threshold, _should_promote_to_steer,
  _promote_drift_to_steer, _compose_goldfive_steer_body)
* _apply_user_steer_state (#199 / #183 plumbing)
* Refine-outcome bookkeeping (_record_refine_outcome,
  reset_for_turn, _occurrence_count_for_ladder,
  _escalate_refine_failure_as_critical_drift)

Audit #402 (dispatch-before-plan-swap) preserved, not fixed: the
GOLDFIVE_STEER ControlMessage dispatch in _promote_drift_to_steer
still happens BEFORE planner.refine_steer + plan install, so the
payload carries the prior plan's task ids. AUDIT #402 comment marks
the preserved ordering. The fix has its own queued PR.

Stays on router:
* _should_inject (reads router-owned _observation_only)
* All 21 shim methods at the steerer.X surface (handle_drift,
  promote, dispatch_*, request_invocation_cancel,
  apply_user_steer_state, record_refine_outcome, reset_for_turn,
  ...) so test fixtures + the planner + executors keep working
  byte-equivalently.
* InterventionLevel enum + RefineExhausted exception (module-level
  exports kept stable for back-compat).
* REFINE_FAILURE_THRESHOLD, PROGRESS_STALL_THRESHOLD_SECONDS class
  constants (subclasses + tests tune at DefaultSteerer.X /
  steerer.X).

Logger contract: kept the bucket-3b sibling _steerer_log pattern.
The migrated INFO/WARNING lines (SKIPPING GOLDFIVE_*, redundant
verdict, no-op revision skipped, planner.refine raised, etc.) still
emit under the "goldfive.steerer" logger name so the six caplog
test files (test_observation_only*, test_judge_task_lifetime,
test_reasoning_judge, test_user_steer_invariant) plus harmonograf /
operator log dashboards keep working with zero churn. Pure debug
lines unique to the new home use the "goldfive.drift_observer"
logger.

LOC: steerer.py 4100 -> 2050 (target was ~2300, came in under);
drift_observer.py 2345 -> 4571.

Test results: 2540 passed, 59 skipped (pre-existing), 0 failures.
Strict state-ownership audit (GOLDFIVE_STRICT_STATE_OWNERSHIP=1)
also clean: 2540 passed.

Fragile-scenario traces verified:
* CRITICAL OFF_TOPIC -> observe_reasoning -> handle_drift ->
  promote_to_steer -> dispatch_steer_control with stale ids
  (audit #402 ordering survives).
* Background judge LOOPING_REASONING -> late-drift gate
  (active_invocation_ids() / cancel-requested) -> handle_drift
  ladder.
* USER_STEER -> handle_drift -> apply_user_steer_state plumbing
  + plan install via plan_reviser (out of bucket 3c scope).
