+++ goldfive/drift_observer.py
@@ -1,10 +1,15 @@
-"""Drift event lifecycle + classification + observation helpers for :class:`DefaultSteerer`.
+"""Drift event lifecycle + classification + observation + dispatch helpers for :class:`DefaultSteerer`.
 
 Extracted from :mod:`goldfive.steerer` in Wave C of the steerer split.
-Buckets **3a** (observability primitives) and **3b** (observation entry
-points + judge orchestration) now live here; bucket **3c** (dispatch +
-ladder + promotion) remains on :class:`DefaultSteerer` and is the next
-follow-up PR.
+Buckets **3a** (observability primitives), **3b** (observation entry
+points + judge orchestration), and **3c** (dispatch + ladder + promotion)
+all now live here. :class:`DefaultSteerer` is the thin router that owns
+the shared mutable state (sinks, planner, adapter, control_channel,
+ContextVar, background-task sets, plan locks, ``observation_only`` flag,
+config, ``REFINE_FAILURE_THRESHOLD`` / ``PROGRESS_STALL_THRESHOLD_SECONDS``
+class constants) plus the public shim layer that tests + external
+callers (the planner, executors) historically hooked at the
+``steerer.X`` bare-attribute name.
 
 This module owns the drift-event observability + classification surface
 plus the observation entry points (``observe`` / ``observe_reasoning``),
@@ -13,12 +18,23 @@ the background judge orchestration (``_run_judge_background`` /
 reflective self-progress check (``maybe_run_reflective_check`` /
 ``note_llm_call`` / ``_emit_reflective_failure``), the GOAL_DRIFT
 trajectory-level check (``maybe_run_goal_drift_check`` /
-``note_agent_turn`` / ``_maybe_run_goal_drift_on_task_boundary``), and
-the bounded note-buffer family (``note_agent_activity`` /
-``note_tool_observation``). The dispatch + ladder + promotion machinery
-(``_handle_drift`` / ``_promote_drift_to_steer`` / the intervention
-ladder / cancel-inflight / late-drift gate) stays on
-:class:`DefaultSteerer` for now — that is bucket 3c.
+``note_agent_turn`` / ``_maybe_run_goal_drift_on_task_boundary``), the
+bounded note-buffer family (``note_agent_activity`` /
+``note_tool_observation``), and the **dispatch + ladder + promotion**
+machinery: ``handle_drift`` (the central drift-routing method),
+``_promote_drift_to_steer`` (audit issue #402 — dispatch-before-plan-swap
+ordering is preserved here, not fixed), the intervention ladder
+(``_ladder_level_for`` / ``_LADDER`` / ``_LADDER_BY_VALUE``), ladder
+dispatch (``_dispatch_nudge`` / ``_dispatch_goldfive_steer_control`` /
+``_dispatch_goldfive_pause_control`` / ``_dispatch_pause_escalate``),
+adapter cancel tagging, the late-drift gate
+(``_is_late_drift_for_terminated_invocation``),
+``request_invocation_cancel``, ``_cancel_inflight_for_revision``,
+``_apply_user_steer_state`` (#199 / #183 synthetic-USER_STEER
+suppression plumbing), and refine-outcome bookkeeping
+(``_record_refine_outcome`` / ``reset_for_turn`` /
+``_occurrence_count_for_ladder`` /
+``_escalate_refine_failure_as_critical_drift``).
 
 Responsibilities (this PR)
 --------------------------
@@ -78,31 +94,34 @@ Responsibilities (this PR)
 * :meth:`_parse_reflective_response` — tolerant JSON-from-LLM parser
   used by the reflective check verdict path.
 
-The module DOES NOT own (yet — follow-up PR)
---------------------------------------------
+The module DOES NOT own
+----------------------
 
-* :meth:`_handle_drift` (753 lines), the intervention ladder
-  (:meth:`_ladder_level_for`, :attr:`_LADDER` / :attr:`_LADDER_BY_VALUE`),
-  ladder dispatch (``_dispatch_nudge`` / ``_dispatch_goldfive_steer_control``
-  / ``_dispatch_goldfive_pause_control`` / ``_dispatch_pause_escalate``),
-  adapter cancel tagging, late-drift gate
-  (``_is_late_drift_for_terminated_invocation``),
-  ``request_invocation_cancel``, ``_cancel_inflight_for_revision``,
-  promotion (``_should_promote_to_steer``, ``_promote_drift_to_steer``,
-  ``_compose_goldfive_steer_body``), ``_apply_user_steer_state``,
-  refine-outcome bookkeeping (``_record_refine_outcome`` /
-  ``reset_for_turn`` / ``_occurrence_count_for_ladder`` /
-  ``_escalate_refine_failure_as_critical_drift``), structural
-  escalation (``_is_task_progress_stalled`` /
-  ``_emit_progress_stalled_escalation`` /
-  ``_emit_handler_exhausted_escalation``). Moved in **bucket 3c** —
-  the largest and most fragile cut, kept separate so audit issues
-  #402 (dispatch-before-plan-swap) and #403 (lock window) can be
-  reviewed against a single focused diff.
 * Task-status transitions — moved in bucket 1
   (:mod:`goldfive.task_state_machine`).
 * Plan-revision install + refine observability — moved in bucket 2
   (:mod:`goldfive.plan_reviser`).
+* Class constants ``REFINE_FAILURE_THRESHOLD`` and
+  ``PROGRESS_STALL_THRESHOLD_SECONDS`` — stay on
+  :class:`DefaultSteerer` so subclasses + tests that historically
+  tuned them at ``DefaultSteerer.X`` / ``steerer.X`` keep working;
+  read from this module as ``self._steerer.X``.
+* The :class:`~goldfive.steerer.InterventionLevel` enum — stays as a
+  module-level export of :mod:`goldfive.steerer` for back-compat with
+  callers that import it as ``from goldfive.steerer import
+  InterventionLevel``.
+
+Audit issue #402 — dispatch-before-plan-swap (preserved, not fixed)
+-------------------------------------------------------------------
+
+:meth:`_promote_drift_to_steer` dispatches the ``GOLDFIVE_STEER``
+ControlMessage **before** calling ``planner.refine_steer`` and
+installing the resulting plan. The dispatch payload therefore carries
+the *prior* plan's task ids in ``superseded_task_ids`` /
+``replacement_task_ids`` (audit issue #402 — the executor sees stale
+ids in the channel message). Bucket 3c preserves this ordering
+verbatim; the fix lives in a separate queued PR so the refactor diff
+and the bug fix stay reviewable.
 
 All cross-component calls go through the router back-reference passed
 to :meth:`DriftObserver.__init__` (``self._steerer``). This keeps the
@@ -114,13 +133,15 @@ components decoupled and lets the router own the shared state
 from __future__ import annotations
 
 import asyncio
+import inspect
 import json
 import logging
 import re
 import time
-from collections.abc import Awaitable, Callable
+from collections.abc import Awaitable, Callable, Mapping
 from typing import TYPE_CHECKING, Any
 
+from goldfive import _state_audit
 from goldfive import state_store as _ostate
 from goldfive.drift import (
     classify_refusal,
@@ -128,15 +149,19 @@ from goldfive.drift import (
     classify_tool_error,
 )
 from goldfive.types import (
+    CancellationRequest,
     DriftEvent,
     DriftKind,
     DriftSeverity,
+    Plan,
+    RefineOutcome,
     Session,
     Task,
+    TaskStatus,
 )
 
 if TYPE_CHECKING:
-    from goldfive.steerer import DefaultSteerer
+    from goldfive.steerer import DefaultSteerer, InterventionLevel
 
 # Shape of the opt-in reflective LLM callable. Re-exported here so the
 # DriftObserver methods that own the reflective + goal-drift + reasoning
@@ -157,21 +182,23 @@ _steerer_log = logging.getLogger("goldfive.steerer")
 
 
 class DriftObserver:
-    """Drift event observability + classification helpers.
+    """Drift event observability + classification + dispatch helpers.
 
     Constructed by :class:`DefaultSteerer` and held on
     ``DefaultSteerer._drift_observer``. The router delegates the
-    drift-emit / detection / attribution surface to the matching method
-    on this class via thin shims that forward arguments verbatim. Tests
-    that historically poked the bare-attribute names on the steerer
-    (``steerer._emit_drift_detected``, ``steerer._resolve_authored_by``,
-    ``steerer.detect_drift``, etc.) keep working through those shims;
-    nothing on the public surface changes.
+    drift-emit / detection / attribution / dispatch surface to the
+    matching method on this class via thin shims that forward
+    arguments verbatim. Tests that historically poked the
+    bare-attribute names on the steerer
+    (``steerer._emit_drift_detected``, ``steerer._handle_drift``,
+    ``steerer._promote_drift_to_steer``,
+    ``steerer.request_invocation_cancel``, etc.) keep working through
+    those shims; nothing on the public surface changes.
 
-    This is **bucket 3a** of the steerer split — the observability
-    cluster. The dispatch + ladder + promotion machinery is owned by
-    :class:`DefaultSteerer` for now and will migrate in subsequent
-    follow-up PRs (3b, 3c).
+    This is the post-**bucket 3c** state of the steerer split: bucket
+    3a (observability primitives), 3b (observation entry points +
+    judge orchestration), and 3c (dispatch + ladder + promotion +
+    refine-outcome bookkeeping) all live here.
     """
 
     # ------------------------------------------------------------------
@@ -244,11 +271,19 @@ class DriftObserver:
     def __init__(self, steerer: DefaultSteerer) -> None:
         # Back-reference to the router. Used to reach the shared
         # event-emission primitives (``_new_envelope`` / ``_emit`` /
-        # ``_drift_kind_pb_value`` / ``_drift_severity_pb_value``) and
+        # ``_drift_kind_pb_value`` / ``_drift_severity_pb_value``),
         # the cross-component cooperators (the bound adapter for
-        # boundary cleanup, the dispatch / ladder methods that still
-        # live on :class:`DefaultSteerer` and will migrate in
-        # follow-up bucket-3 PRs).
+        # boundary cleanup, the planner, the plan-revision install +
+        # refine-observability surface owned by
+        # :class:`~goldfive.plan_reviser.PlanReviser` and the
+        # task-state-machine surface owned by
+        # :class:`~goldfive.task_state_machine.TaskStateMachine`), and
+        # the router-owned shared state (the per-async-task
+        # ``_active_session_var`` ContextVar plumbing for the planner
+        # span-context provider, the ``_observation_only`` gate read
+        # via :meth:`DefaultSteerer._should_inject`, and the
+        # ``REFINE_FAILURE_THRESHOLD`` / ``PROGRESS_STALL_THRESHOLD_SECONDS``
+        # class constants).
         self._steerer = steerer
 
     # ------------------------------------------------------------------
@@ -701,7 +736,7 @@ class DriftObserver:
             current_task_id=parent_task_id,
             current_agent_id=assignee,
         )
-        await self._steerer._handle_drift(drift, session)
+        await self.handle_drift(drift, session)
 
     async def report_plan_divergence(
         self,
@@ -914,7 +949,7 @@ class DriftObserver:
             f"{age:.0f}s since last progress, threshold "
             f"{self._steerer.PROGRESS_STALL_THRESHOLD_SECONDS:.0f}s"
         )
-        await self._steerer._dispatch_goldfive_pause_control(drift, session, reason=reason)
+        await self._dispatch_goldfive_pause_control(drift, session, reason=reason)
         escalation = DriftEvent(
             kind=DriftKind.HUMAN_INTERVENTION_REQUIRED,
             severity=DriftSeverity.CRITICAL,
@@ -945,7 +980,7 @@ class DriftObserver:
             f"{drift.current_task_id or '(trajectory)'}: "
             f"planner cannot produce a meaningful change"
         )
-        await self._steerer._dispatch_goldfive_pause_control(drift, session, reason=reason)
+        await self._dispatch_goldfive_pause_control(drift, session, reason=reason)
         escalation = DriftEvent(
             kind=DriftKind.HUMAN_INTERVENTION_REQUIRED,
             severity=DriftSeverity.CRITICAL,
@@ -1054,7 +1089,7 @@ class DriftObserver:
             drift = self._steerer.detect_drift(event, session)
         if drift is None:
             return
-        await self._steerer._handle_drift(drift, session)
+        await self.handle_drift(drift, session)
 
     async def observe_reasoning(
         self,
@@ -1127,7 +1162,7 @@ class DriftObserver:
             # framework-agnostic).
             if not drift.trigger_input:
                 drift.trigger_input = self._truncate_trigger_input(text)
-            await self._steerer._handle_drift(drift, session)
+            await self.handle_drift(drift, session)
             return
 
         # Mode-selected pipeline (judge / embedding / both / off).
@@ -1296,7 +1331,7 @@ class DriftObserver:
             # verdicts that may outlive the originating invocation —
             # synchronous detectors run inline on the model-response
             # callback and always see a live invocation.
-            if self._steerer._is_late_drift_for_terminated_invocation(drift, session):
+            if self._is_late_drift_for_terminated_invocation(drift, session):
                 _steerer_log.info(
                     "DefaultSteerer: stale judge verdict; invocation for "
                     "agent=%r task=%r already terminated; drift kind=%s "
@@ -1309,7 +1344,7 @@ class DriftObserver:
                     drift.authored_by = self._resolve_authored_by(drift)
                 await self._emit_drift_detected(session, drift)
                 return
-            await self._steerer._handle_drift(drift, session)
+            await self.handle_drift(drift, session)
         except asyncio.CancelledError:
             # Propagate cancellation so :meth:`shutdown` / event-loop
             # teardown can cleanly abort a still-running judge without
@@ -1846,7 +1881,7 @@ class DriftObserver:
                 current_task_id=task.id,
                 current_agent_id=agent_id_for_drift,
             )
-            await self._steerer._handle_drift(drift, session)
+            await self.handle_drift(drift, session)
             return
         if conf_val < 0.5:
             drift = DriftEvent(
@@ -1859,7 +1894,7 @@ class DriftObserver:
                 current_task_id=task.id,
                 current_agent_id=agent_id_for_drift,
             )
-            await self._steerer._handle_drift(drift, session)
+            await self.handle_drift(drift, session)
             return
         # making_progress=true, confidence >= 0.5 -- no drift.
         return
@@ -2161,7 +2196,7 @@ class DriftObserver:
         )
         if drift is None:
             return
-        await self._steerer._handle_drift(drift, session)
+        await self.handle_drift(drift, session)
 
     def _spawn_goal_drift_judge_background(self, session: Session) -> None:
         """Spawn :meth:`maybe_run_goal_drift_check` as a fire-and-forget task.
@@ -2317,7 +2352,7 @@ class DriftObserver:
         ``WARNING`` muddying the signal.
         """
         try:
-            await self._steerer._handle_drift(drift, session)
+            await self.handle_drift(drift, session)
         except asyncio.CancelledError:
             raise
         except Exception as exc:  # noqa: BLE001 — background task
@@ -2343,3 +2378,2202 @@ class DriftObserver:
         # One yield so the ``add_done_callback(...discard)`` has run
         # and the set is fully empty for the next assertion / spawn.
         await asyncio.sleep(0)
+
+    # ==================================================================
+    # Bucket 3c — Dispatch + intervention ladder + promotion
+    # ==================================================================
+    #
+    # Methods below were extracted verbatim from
+    # :class:`goldfive.steerer.DefaultSteerer` in the bucket-3c step of
+    # the steerer split. ``self.X`` references that target router-level
+    # state (sinks, planner, adapter, control_channel, plan locks,
+    # observation_only flag, REFINE_FAILURE_THRESHOLD /
+    # PROGRESS_STALL_THRESHOLD_SECONDS class constants, the
+    # ContextVar-isolated active-session plumbing) read via
+    # ``self._steerer.X``; ``self.X`` references that target the
+    # observability primitives extracted in buckets 3a/3b (drift emit,
+    # detect, ladder constants, structural escalation helpers) read
+    # directly off this class.
+    #
+    # The most-fixed of the moved methods is :meth:`handle_drift` (the
+    # ex-``_handle_drift``); the line count goes well past terseness
+    # norms because the source did. :meth:`_promote_drift_to_steer`
+    # contains audit issue #402 (dispatch-before-plan-swap) which is
+    # **preserved**, not fixed — the fix has its own queued PR.
+
+    # ------------------------------------------------------------------
+    # Intervention ladder table (goldfive#142)
+    # ------------------------------------------------------------------
+    #
+    # See the long-form rationale comments above
+    # :attr:`goldfive.steerer.DefaultSteerer._LADDER` (now removed from
+    # the router; this is the canonical home post bucket-3c).
+
+    _LADDER: dict[
+        DriftKind,
+        tuple[
+            "InterventionLevel | None",
+            "InterventionLevel | None",
+            tuple["InterventionLevel", "InterventionLevel"],
+        ],
+    ] = {}  # populated lazily in :meth:`_load_ladder_tables` to avoid an
+    # import-cycle with :mod:`goldfive.steerer` (which defines the
+    # :class:`InterventionLevel` enum).
+
+    _LADDER_BY_VALUE: dict[
+        str,
+        tuple[
+            "InterventionLevel | None",
+            "InterventionLevel | None",
+            tuple["InterventionLevel", "InterventionLevel"],
+        ],
+    ] = {}
+
+    _LADDER_LOADED: bool = False
+
+    @classmethod
+    def _load_ladder_tables(cls) -> None:
+        """Populate :attr:`_LADDER` / :attr:`_LADDER_BY_VALUE` on first use.
+
+        Lazy load defers the :class:`InterventionLevel` import so this
+        module can be imported before :mod:`goldfive.steerer` finishes
+        its own imports (the steerer constructs a :class:`DriftObserver`
+        in its ``__init__``).
+        """
+        if cls._LADDER_LOADED:
+            return
+        from goldfive.steerer import InterventionLevel as _IL
+
+        cls._LADDER = {
+            DriftKind.CONFABULATION_RISK: (
+                _IL.OBSERVE,
+                _IL.ABSORB,
+                (_IL.CANCEL_REINVOKE, _IL.PAUSE_ESCALATE),
+            ),
+            DriftKind.AGENT_REFUSAL: (
+                _IL.OBSERVE,
+                _IL.ABSORB,
+                (_IL.CANCEL_REINVOKE, _IL.PAUSE_ESCALATE),
+            ),
+            DriftKind.MODEL_REFUSAL: (
+                _IL.OBSERVE,
+                _IL.ABSORB,
+                (_IL.CANCEL_REINVOKE, _IL.PAUSE_ESCALATE),
+            ),
+            DriftKind.LOOPING_REASONING: (
+                None,
+                _IL.ABSORB,
+                (_IL.NUDGE, _IL.PAUSE_ESCALATE),
+            ),
+            DriftKind.LOOPING_TOOL_CALL: (
+                None,
+                _IL.ABSORB,
+                (_IL.CANCEL_REINVOKE, _IL.PAUSE_ESCALATE),
+            ),
+            DriftKind.REASONING_CLUSTER_TIGHTENING: (
+                _IL.OBSERVE,
+                None,
+                (_IL.OBSERVE, _IL.OBSERVE),
+            ),
+            DriftKind.PLAN_DIVERGENCE: (
+                _IL.OBSERVE,
+                _IL.ABSORB,
+                (_IL.CANCEL_REINVOKE, _IL.PAUSE_ESCALATE),
+            ),
+            DriftKind.OFF_TOPIC: (
+                _IL.OBSERVE,
+                _IL.ABSORB,
+                (_IL.CANCEL_REINVOKE, _IL.PAUSE_ESCALATE),
+            ),
+            DriftKind.JUSTIFIED_DEVIATION: (
+                _IL.OBSERVE,
+                _IL.ABSORB,
+                (_IL.ABSORB, _IL.ABSORB),
+            ),
+            DriftKind.INTENT_DIVERGENCE: (
+                _IL.OBSERVE,
+                _IL.ABSORB,
+                (_IL.PAUSE_ESCALATE, _IL.PAUSE_ESCALATE),
+            ),
+            DriftKind.TOOL_ERROR: (
+                _IL.OBSERVE,
+                _IL.ABSORB,
+                (_IL.CANCEL_REINVOKE, _IL.PAUSE_ESCALATE),
+            ),
+            DriftKind.RUNAWAY_DELEGATION: (
+                None,
+                None,
+                (_IL.CANCEL_REINVOKE, _IL.PAUSE_ESCALATE),
+            ),
+            DriftKind.REFINE_VALIDATION_FAILED: (
+                None,
+                None,
+                (_IL.PAUSE_ESCALATE, _IL.PAUSE_ESCALATE),
+            ),
+            DriftKind.HUMAN_INTERVENTION_REQUIRED: (
+                None,
+                None,
+                (_IL.PAUSE_ESCALATE, _IL.TERMINATE),
+            ),
+            DriftKind.GOAL_DRIFT: (
+                None,
+                _IL.NUDGE,
+                (_IL.NUDGE, _IL.CANCEL_REINVOKE),
+            ),
+            DriftKind.SELF_REPORTED_STUCK: (
+                None,
+                _IL.ABSORB,
+                (_IL.CANCEL_REINVOKE, _IL.PAUSE_ESCALATE),
+            ),
+        }
+        cls._LADDER_LOADED = True
+
+    def _ladder_level_for(
+        self,
+        kind: DriftKind,
+        severity: DriftSeverity,
+        occurrence_count: int,
+    ) -> "InterventionLevel":
+        """Return the intervention level for ``(kind, severity, count)``.
+
+        See :meth:`goldfive.steerer.DefaultSteerer._ladder_level_for` —
+        identical semantics; lifted verbatim into bucket 3c so the
+        whole drift-dispatch surface has one home.
+        """
+        from goldfive.steerer import InterventionLevel as _IL
+
+        self._load_ladder_tables()
+        entry = self._LADDER_BY_VALUE.get(kind.value)
+        if entry is None:
+            entry = self._LADDER.get(kind)
+        is_repeat = occurrence_count >= self._steerer.REFINE_FAILURE_THRESHOLD
+        if entry is not None:
+            info_level, warning_level, critical_pair = entry
+            if severity is DriftSeverity.INFO:
+                return info_level or _IL.OBSERVE
+            if severity is DriftSeverity.WARNING:
+                return warning_level or _IL.OBSERVE
+            return critical_pair[1] if is_repeat else critical_pair[0]
+        if severity is DriftSeverity.INFO:
+            return _IL.OBSERVE
+        if severity is DriftSeverity.WARNING:
+            return _IL.ABSORB
+        return _IL.PAUSE_ESCALATE if is_repeat else _IL.ABSORB
+
+    # ------------------------------------------------------------------
+    # The central drift-routing entry point.
+    # ------------------------------------------------------------------
+
+    async def handle_drift(self, drift: DriftEvent, session: Session) -> None:
+        """Emit a ``DriftDetected`` event and dispatch via the intervention ladder.
+
+        Verbatim move of :meth:`DefaultSteerer._handle_drift` (the
+        single most-fixed method on the codebase after
+        :meth:`_emit_plan_revised`). See the original docstring in
+        :class:`DefaultSteerer` for the long-form contract; the body
+        and behaviour are unchanged here.
+        """
+        from goldfive.steerer import (
+            InterventionLevel,
+            RefineExhausted,
+            _ABSORB_NUDGE_KINDS,
+            _planner_refine_accepts_available_agents,
+            compose_corrective_user_message,
+        )
+
+        # goldfive#252: PLAN_DIVERGENCE replaced by CAPABILITY_MISMATCH
+        # (#253) — disabled here. Guard at the very top so any external
+        # producer (legacy callers, replays, sinks) cannot revive it.
+        if drift.kind is DriftKind.PLAN_DIVERGENCE:
+            log.debug(
+                "DefaultSteerer._handle_drift: PLAN_DIVERGENCE drift "
+                "received but handling is disabled (goldfive#252); "
+                "detail=%r",
+                drift.detail,
+            )
+            return
+        # Normalise source attribution early so every downstream
+        # consumer (sinks, promotion policy, prompt framing) sees a
+        # non-empty ``authored_by``. USER_* kinds → "user"; anything
+        # else → "goldfive". Honours an explicit non-empty value on
+        # the drift (e.g. callers that already attributed) via
+        # :meth:`_resolve_authored_by`.
+        if not drift.authored_by:
+            drift.authored_by = self._resolve_authored_by(drift)
+        # goldfive#245 — verdict-freshness gate. Every observation/
+        # detector stamps ``observed_revision_index`` from
+        # ``session.plan.revision_index`` BEFORE its LLM await; the
+        # reconciler may transition tasks during that round-trip, so a
+        # verdict that arrives after the framework moved on is moot.
+        # Drop it here: emit ``DriftDetected`` for observability
+        # (operators see the detector ran) and skip the cancel + refine
+        # machinery.
+        #
+        # Bypasses:
+        #   * Unstamped drifts (``observed_revision_index == 0``) are
+        #     legacy / external producers / pre-#245 emit paths — flow
+        #     through unchanged so the gate is purely additive.
+        #   * User-authored drifts (USER_STEER / USER_CANCEL /
+        #     USER_PAUSE) bypass the gate even when stamped: an
+        #     operator directive must be honoured regardless of the
+        #     framework's plan-state cursor (preserves the iter-11D /
+        #     #242 contract).
+        if drift.observed_revision_index and (drift.authored_by or "").lower() != "user":
+            # Per-(kind, target) addressed-watermark check — narrower
+            # than naive ``observed < live_revision`` gating so parallel
+            # judges firing on *orthogonal* concerns aren't over-rejected.
+            # Naive gating drops a GOAL_DRIFT verdict observed at N just
+            # because an unrelated OFF_TOPIC refine bumped the plan to
+            # N+1; this gate drops only when the SAME (kind, target) was
+            # already addressed at a later revision (genuinely redundant).
+            #
+            # ``last_addressed_revision_by_drift_key`` is stamped by
+            # :meth:`_apply_revision` after every successful goldfive-
+            # authored refine. Empty target (``""``) coalesces trajectory-
+            # level drifts on one key, so trajectory-wide addressing
+            # works correctly.
+            key = (drift.kind.value, drift.current_task_id or "")
+            last_addressed = int(
+                session.last_addressed_revision_by_drift_key.get(key, 0)
+            )
+            if last_addressed and drift.observed_revision_index < last_addressed:
+                _steerer_log.info(
+                    "DefaultSteerer._handle_drift: redundant verdict — "
+                    "drift kind=%s target=%r observed revision %d but "
+                    "same (kind, target) was already addressed at "
+                    "revision %d; skipping dispatch",
+                    drift.kind.value,
+                    drift.current_task_id or "<trajectory>",
+                    drift.observed_revision_index,
+                    last_addressed,
+                )
+                # Emit for observability so operators see the detector
+                # ran; do NOT cancel / refine on a redundant view.
+                await self._emit_drift_detected(session, drift)
+                return
+        # Tag the bound adapter's next cancel with a symbolic reason so
+        # the synthetic function_response the adapter appends on cancel
+        # carries LLM-actionable content. Done BEFORE the drift event
+        # is emitted so a sink that reacts by cancelling the invoke
+        # sees the tag. Harmless if the adapter doesn't expose the
+        # attribute (duck-typed) or no adapter is bound. See
+        # goldfive#139 and
+        # :func:`goldfive.adapters.adk._build_cancelled_response_event`.
+        self._tag_adapter_cancel_reason(drift, session=session)
+        # goldfive#152: USER_STEER-specific side effects -- write the
+        # active-steer bookkeeping onto the orchestration-state dict
+        # and synthesize a durable Goal from the steer body so
+        # subsequent refines see the pivot as a first-class goal,
+        # not a one-shot user message. Done BEFORE the drift event
+        # is emitted (the state writes are cheap) and BEFORE the
+        # ladder dispatches to planner.refine (which reads
+        # ``session.goals`` we just mutated) so the refine sees the
+        # new goal shape in the same dispatch.
+        if drift.kind is DriftKind.USER_STEER:
+            # Plumb the session into the planner's span-context provider
+            # for the duration of synthesize_goal_from_steer so its LLM
+            # call shows up as a span on the Gantt. Cleared in a
+            # ``finally`` so exceptions don't leave a stale pointer.
+            _token = self._steerer._active_session_var.set(session)
+            try:
+                await self._apply_user_steer_state(drift, session)
+            finally:
+                self._steerer._active_session_var.reset(_token)
+        # goldfive-steer-unification: consult the severity-aware
+        # promotion policy BEFORE emitting DriftDetected so that a
+        # suppressed goldfive steer carries the ``suppressed_by_user_steer``
+        # flag on the wire (sinks can surface the suppression
+        # decision). ``_should_promote_to_steer`` returns ``True`` iff
+        # the drift is goldfive-authored, clears the configured
+        # severity threshold, and is not blocked by an active fresh
+        # user steer; as a side effect it stamps
+        # ``drift.suppressed_by_user_steer=True`` when the suppression
+        # path wins.
+        promote_to_steer = self._should_promote_to_steer(drift, session)
+        await self._emit_drift_detected(session, drift)
+        if drift.suppressed_by_user_steer:
+            # Suppression path: the goldfive drift fired, was observed
+            # via DriftDetected, and — per the fresh user-steer
+            # suppression window — we neither cancel nor refine. The
+            # pre-unification passive ladder dispatch is also skipped:
+            # a user steer is already active, its refine has already
+            # happened, and running another refine for this signal
+            # would race against it.
+            return
+        # Cooperative cancellation (goldfive#251 Stream C / 7a). Severity
+        # ladder decision: CRITICAL drifts (and ONLY critical drifts)
+        # flag the currently-active invocation(s) for cooperative
+        # cancel before the refine / promote path runs. INFO + WARNING
+        # severities do NOT cancel — they flow through the usual
+        # observe / absorb / nudge channels. User-authored drifts
+        # (USER_STEER / USER_CANCEL / USER_PAUSE) additionally bypass
+        # the severity gate because an operator directive must be
+        # honoured even when emitted at a lower severity tier.
+        #
+        # The actual short-circuit happens in the ADK plugin's next
+        # ``before_agent_callback`` / ``before_model_callback`` /
+        # ``before_tool_callback``; this call just writes the flag.
+        # Whether to re-dispatch after the cancel is the parent
+        # agent's decision, informed by plan-causal prompting from
+        # Stream B — the framework itself does NOT auto-reinvoke.
+        if self._should_request_cancel_for_drift(drift):
+            try:
+                await self.request_invocation_cancel(drift=drift, session=session)
+            except Exception as exc:  # noqa: BLE001 — cancel is best-effort
+                log.debug(
+                    "DefaultSteerer._handle_drift: request_invocation_cancel raised: %s",
+                    exc,
+                )
+        if promote_to_steer:
+            await self._promote_drift_to_steer(drift, session)
+            return
+        # Route through the intervention ladder. The per-(kind, task)
+        # occurrence count drives the "first vs repeat" distinction in
+        # the ladder table -- we read it BEFORE any mutation so the
+        # mapping sees the state at drift-fire time. ``occurrence_count``
+        # is derived from ``session.refine_outcomes`` via
+        # :meth:`_occurrence_count_for_ladder` (goldfive#215 iter-8 P2:
+        # the outcome dict replaces the deleted ``refine_failure_counts``
+        # int counter).
+        occurrence_count = self._occurrence_count_for_ladder(session, drift)
+        level = self._ladder_level_for(drift.kind, drift.severity, occurrence_count)
+        log.debug(
+            "DefaultSteerer._handle_drift: kind=%s severity=%s occurrence=%d -> level=%s",
+            drift.kind.value,
+            drift.severity.value,
+            occurrence_count,
+            level.name,
+        )
+        if level is InterventionLevel.OBSERVE:
+            return
+        if level is InterventionLevel.NUDGE:
+            await self._dispatch_nudge(drift, session)
+            return
+        if level is InterventionLevel.PAUSE_ESCALATE:
+            await self._dispatch_pause_escalate(drift, session)
+            return
+        if level is InterventionLevel.TERMINATE:
+            # Level 5 is reserved for a future Runner-side timeout on a
+            # stuck Level 4 pause. Today we fall back to PAUSE_ESCALATE
+            # so no code path silently drops the drift.
+            await self._dispatch_pause_escalate(drift, session)
+            return
+        # ABSORB and CANCEL_REINVOKE both call ``planner.refine`` and
+        # install the revised plan. CANCEL_REINVOKE additionally queues
+        # a corrective message on the session for the overlay loop
+        # (goldfive#141). The refine call itself is identical so we
+        # share the implementation below and read the level at the end
+        # to decide whether to emit the follow-up handoff.
+        if self._steerer._planner is None or session.plan is None:
+            return
+        if drift.kind is DriftKind.REFINE_VALIDATION_FAILED:
+            # Terminal planner signal (goldfive#133). Do NOT call refine
+            # again on it. The ladder already routes this to Level 4 so
+            # control flow normally won't reach here, but belt-and-braces.
+            return
+        # Outcome-based gate (goldfive#215 iter-8 P2 — unified G1+G3).
+        # Skip refine when ``(kind, task)`` already has a terminal
+        # outcome on this turn: a prior refine already succeeded (the
+        # current drift is a same-turn replay of an addressed
+        # condition), or prior refines have failed
+        # >= REFINE_FAILURE_THRESHOLD times (the threshold trip already
+        # marked the task FAILED + emitted REPEATED_FAILURE; a third
+        # tick must not retry). USER_STEER / USER_CANCEL / GOAL_DRIFT
+        # bypass the gate — operator intent always honoured,
+        # trajectory drifts have their own rate limiters.
+        if drift.kind not in self._USER_OR_TRAJECTORY_DRIFT_KINDS:
+            outcome_key = (drift.kind.value, drift.current_task_id or "")
+            outcome = session.refine_outcomes.get(outcome_key)
+            if outcome is not None:
+                if outcome.state == "succeeded":
+                    log.debug(
+                        "refine skipped: prior succeeded outcome (kind=%s task=%r)",
+                        drift.kind.value,
+                        drift.current_task_id,
+                    )
+                    return
+                if outcome.fail_count >= self._steerer.REFINE_FAILURE_THRESHOLD:
+                    log.debug(
+                        "refine skipped: failure threshold reached (kind=%s task=%r count=%d)",
+                        drift.kind.value,
+                        drift.current_task_id,
+                        outcome.fail_count,
+                    )
+                    return
+        # Progress-based escalation (goldfive#271). Orthogonal to the
+        # outcome gate: a task that has been silent past the configured
+        # stall threshold escalates to HUMAN_INTERVENTION_REQUIRED
+        # instead of looping the planner. A productively-iterating task
+        # has continuous progress events; a stuck task does not.
+        if self._is_task_progress_stalled(drift, session):
+            await self._emit_progress_stalled_escalation(drift, session)
+            return
+        # Plumb the session into the planner's drift-emitter callback
+        # for the duration of this refine call so the planner can emit
+        # REFINE_VALIDATION_FAILED drifts through the normal event
+        # pipeline. Cleared in a ``finally`` so exceptions don't leave
+        # a stale session pointer. ContextVar isolation keeps concurrent
+        # runs from stomping each other (goldfive#133, PR #294 audit).
+        _active_session_token = self._steerer._active_session_var.set(session)
+        # goldfive a4: mint a refine-attempt id for correlation across
+        # ``refine_attempted`` and the paired success/failure event.
+        attempt_id = self._steerer._new_attempt_id()
+        await self._steerer._emit_refine_attempted(session, drift, attempt_id=attempt_id)
+        try:
+            # Thread the adapter's available_agents_tree (goldfive#151)
+            # through refine so the LLM is constrained to pick real
+            # tree assignees. Adapters without the property fall back
+            # to ``available_agents`` (list[str]); custom/legacy adapters
+            # without either surface produce ``None`` and the planner
+            # keeps its pre-#151 behaviour. Planners whose refine does
+            # not accept the kwarg (test stubs, pre-#151 custom
+            # planners) are called the old way so nothing breaks.
+            available_agents: Any = None
+            adapter = self._steerer._adapter
+            if adapter is not None:
+                tree = getattr(adapter, "available_agents_tree", None)
+                if isinstance(tree, list) and tree:
+                    available_agents = list(tree)
+                else:
+                    flat = getattr(adapter, "available_agents", None)
+                    if flat:
+                        available_agents = list(flat)
+            refine_accepts_registry = _planner_refine_accepts_available_agents(
+                self._steerer._planner
+            )
+            # Phase 3.5 (goldfive#271) tripwire wrapper — the
+            # ``except BaseException: stash; raise`` arm below is the
+            # compliance branch (CANCELLATION-CONTRACT.md §1.2). The
+            # boundary catch site at ``ADKAdapter._invoke_internal``
+            # asserts ``mark_stash_completed()`` fired before the
+            # cancel propagated past us.
+            with _state_audit.cancellation_stash_audited("DefaultSteerer._handle_drift.refine"):
+                try:
+                    if refine_accepts_registry:
+                        revised = await self._steerer._planner.refine(
+                            plan=session.plan,
+                            drift=drift,
+                            goals=list(session.goals),
+                            available_agents=available_agents,
+                        )
+                    else:
+                        revised = await self._steerer._planner.refine(
+                            plan=session.plan,
+                            drift=drift,
+                            goals=list(session.goals),
+                        )
+                except RefineExhausted as exc:
+                    # goldfive#271: planner explicitly signals it cannot
+                    # produce a meaningful change. Same escalation path
+                    # as the structural no-op detector — pause for
+                    # human intervention rather than retrying.
+                    _steerer_log.info(
+                        "DefaultSteerer._handle_drift: planner.refine raised "
+                        "RefineExhausted for kind=%s task=%r: %s",
+                        drift.kind.value,
+                        drift.current_task_id,
+                        exc,
+                    )
+                    await self._steerer._emit_refine_failed(
+                        session,
+                        drift,
+                        attempt_id=attempt_id,
+                        failure_kind="refine_exhausted",
+                        reason=str(exc) or "planner signalled handler exhaustion",
+                        detail="",
+                    )
+                    await self._emit_handler_exhausted_escalation(drift, session)
+                    return
+                except Exception as exc:  # noqa: BLE001 — refine errors must not break the run
+                    # Surface the failure via logging + a synthetic follow-up
+                    # drift so operators don't silently see the same plan loop
+                    # forever. Without this, a refine that raises (e.g. malformed
+                    # LLM JSON after a mid-invocation cancel poisons the session)
+                    # leaves session.plan unchanged and the executor re-enters
+                    # the same state on the next tick.
+                    _steerer_log.warning(
+                        "DefaultSteerer._handle_drift: planner.refine(kind=%s) raised "
+                        "%s; plan unchanged",
+                        drift.kind.value,
+                        exc,
+                    )
+                    await self._steerer._emit_refine_failed(
+                        session,
+                        drift,
+                        attempt_id=attempt_id,
+                        failure_kind="llm_error",
+                        reason=str(exc),
+                        detail=type(exc).__name__,
+                    )
+                    await self._escalate_refine_failure_as_critical_drift(
+                        session, drift, reason=str(exc)
+                    )
+                    await self._record_refine_outcome(session, drift, succeeded=False)
+                    return
+                except BaseException as exc:  # noqa: BLE001
+                    # Phase 3.5 (CANCELLATION-CONTRACT.md §C4): ``CancelledError``
+                    # bypasses the ``except Exception`` branch (it is a
+                    # ``BaseException`` since Py 3.8). Emit the paired
+                    # ``refine_failed`` observability event so a refine cancelled
+                    # mid-flight does not leave sinks with an unmatched
+                    # ``refine_attempted``. The ``finally`` below still resets
+                    # ``_active_session_var``; we only own the paired-event
+                    # stash here. Re-raise so cancellation continues to
+                    # propagate per the asyncio contract.
+                    await self._steerer._emit_refine_failed(
+                        session,
+                        drift,
+                        attempt_id=attempt_id,
+                        failure_kind="cancelled",
+                        reason=f"refine cancelled: {type(exc).__name__}",
+                        detail=type(exc).__name__,
+                    )
+                    # Phase 3.5 tripwire compliance marker (§1.2 form).
+                    _state_audit.mark_stash_completed()
+                    raise
+        finally:
+            self._steerer._active_session_var.reset(_active_session_token)
+        if revised is None:
+            # iter-12 (#204): refine returning None at the steerer level
+            # means the planner has already exhausted its internal retry
+            # budget (iter-11C's repeat-rejection guard). Treat as
+            # handler exhaustion and pause for human intervention rather
+            # than emitting a follow-up CRITICAL drift that would
+            # recurse through ``_handle_drift`` and eventually abort the
+            # run. Mirrors the ``RefineExhausted`` and no-op-revision
+            # escalation paths.
+            _steerer_log.warning(
+                "DefaultSteerer._handle_drift: planner.refine(kind=%s) returned None; "
+                "plan unchanged — escalating to HUMAN_INTERVENTION_REQUIRED",
+                drift.kind.value,
+            )
+            await self._steerer._emit_refine_failed(
+                session,
+                drift,
+                attempt_id=attempt_id,
+                failure_kind="parse_error",
+                reason="planner returned no revised plan",
+                detail="",
+            )
+            await self._record_refine_outcome(session, drift, succeeded=False)
+            await self._emit_handler_exhausted_escalation(drift, session)
+            return
+        # I4 fix: fold runtime terminal statuses from the prior plan
+        # onto the revised plan BEFORE validation. A task that was
+        # cancelled / failed / NOT_NEEDED out-of-band between revisions
+        # (e.g. overlay reap → NOT_NEEDED, executor reachability audit
+        # → CANCELLED, coordinator reporting-tool → COMPLETED) should
+        # carry that status into the persisted snapshot, even when the
+        # LLM's view of the prior plan was stale.
+        # goldfive#247: fold returns a NEW Plan (Plan is frozen).
+        revised = self._steerer._fold_runtime_terminal_statuses(revised, session.plan)
+        try:
+            revised.validate(for_revision=True, prior=session.plan)
+        except ValueError as exc:
+            # iter-12 (#204): the revised plan is structurally invalid
+            # AND the planner has already exhausted its internal
+            # validator-rejection retry budget (iter-11C). Treat as
+            # handler exhaustion and pause for human intervention.
+            #
+            # Operator visibility: the SCHEMA_VIOLATION drift is
+            # preserved at INFO severity (observability-only — does NOT
+            # recurse through ``_handle_drift``) so harmonograf and
+            # other sinks still see the schema-failure signal carrying
+            # the validator's reason. The actionable signal is the
+            # paired ``refine_failed(validator_rejected)`` envelope and
+            # the HUMAN_INTERVENTION_REQUIRED escalation that follows.
+            #
+            # Passing ``prior=session.plan`` to ``validate`` enables
+            # PLAN-LIFECYCLE.md §3.1 (terminal task preservation) and
+            # §3.2 (terminal->terminal edge preservation) on top of the
+            # usual structural checks.
+            await self._steerer._emit_refine_failed(
+                session,
+                drift,
+                attempt_id=attempt_id,
+                failure_kind="validator_rejected",
+                reason=f"plan validation failed: {exc}",
+                detail=type(exc).__name__,
+            )
+            await self._emit_drift_detected(
+                session,
+                DriftEvent(
+                    kind=DriftKind.SCHEMA_VIOLATION,
+                    severity=DriftSeverity.INFO,
+                    detail=f"plan validation failed: {exc}",
+                    current_task_id=session.current_task_id,
+                ),
+            )
+            await self._record_refine_outcome(session, drift, succeeded=False)
+            await self._emit_handler_exhausted_escalation(drift, session)
+            return
+        # No-op revision rejection (goldfive#271 — replaces the deleted
+        # count cap). If the LLM produced a "refine" that is structurally
+        # identical to the prior plan (same task ids, edges, assignees,
+        # statuses), treat the handler as exhausted: the planner cannot
+        # produce a meaningful change for this drift. Escalate to
+        # HUMAN_INTERVENTION_REQUIRED rather than bumping the revision
+        # index for a no-op, which would otherwise loop forever on a
+        # judge that keeps re-firing on a corrected task.
+        if self._steerer._plans_structurally_identical(session.plan, revised):
+            _steerer_log.info(
+                "no-op revision skipped (kind=%s task=%r); escalating to "
+                "HUMAN_INTERVENTION_REQUIRED",
+                drift.kind.value,
+                drift.current_task_id,
+            )
+            await self._steerer._emit_refine_failed(
+                session,
+                drift,
+                attempt_id=attempt_id,
+                failure_kind="no_op_revision",
+                reason="planner returned structurally identical plan",
+                detail="",
+            )
+            await self._emit_handler_exhausted_escalation(drift, session)
+            return
+        # Successful refine — record the "succeeded" outcome so a
+        # follow-up same-(kind, task) drift on this turn skips refine
+        # (the prior refine already addressed it).
+        await self._record_refine_outcome(session, drift, succeeded=True)
+        # Capture the outgoing plan BEFORE _apply_revision installs the
+        # revised one; _emit_plan_revised diffs the two to populate the
+        # PlanRevisionDiff sidecar (PLAN-LIFECYCLE.md §2, §8 gap #4).
+        prev_plan = session.plan
+        # goldfive#247: _apply_revision returns the stamped instance.
+        # goldfive#255: _apply_revision returns ``(revised, was_installed)``
+        # so the caller can thread the install outcome into PlanRevised's
+        # ``dry_run`` marker.
+        revised, was_installed = self._steerer._apply_revision(session, revised, drift)
+        # Cancel the in-flight coordinator invocation now that the plan
+        # it was reasoning against has been superseded (goldfive#271
+        # follow-up — v15 concurrent-invocation bug). Order: cancel
+        # BEFORE PlanRevised emit so the synthetic InvocationCancelled
+        # sink event lands adjacent to the revision in the wire log
+        # and operators can correlate the two. Best-effort, never
+        # raises — a no-op cancel still leaves the new plan installed.
+        await self._cancel_inflight_for_revision(drift, session)
+        await self._steerer._emit_plan_revised(
+            session,
+            revised,
+            drift,
+            prev_plan=prev_plan,
+            attempt_id=attempt_id,
+            dry_run=not was_installed,
+        )
+        # Level 3 (CANCEL_REINVOKE) handoff (Phase 2 of the path-
+        # duality fix). Pre-Phase-2 this stuffed
+        # ``session.pending_corrective_message`` — a write-only slot
+        # nobody read after the overlay loop took shape, leaving the
+        # coordinator running its original chain blind to the plan
+        # swap. Phase 2 dispatches a ``GOLDFIVE_STEER`` ControlMessage
+        # so the executor's invoke loop cancels in-flight work and
+        # restarts with the corrective body framed as ``[GOLDFIVE
+        # STEERING CONTROL …]`` — the same junction USER_STEER uses.
+        if level is InterventionLevel.CANCEL_REINVOKE:
+            await self._dispatch_goldfive_steer_control(drift, session)
+        # goldfive#202: for drifts where the coordinator has no way to
+        # observe the plan revision on its own (it is still mid-
+        # invocation, retrying the superseded task), ALSO queue a
+        # Level 2 nudge after a successful ABSORB. The overlay loop's
+        # scoped nudge-replay path (see SequentialExecutor._run_overlay)
+        # picks this up at invocation end and re-invokes the
+        # passthrough with the nudge as the next user message — the
+        # only way for the coordinator to learn its plan changed.
+        #
+        # Scoped to drift kinds whose mid-invocation signature is
+        # "coordinator is stuck on a task goldfive just replaced":
+        # LOOPING_REASONING / LOOPING_TOOL_CALL (detector fires while
+        # the coordinator retries the same tool call), SELF_REPORTED_STUCK
+        # (reflective self-check reports no progress). Other ABSORB
+        # kinds (CONFABULATION_RISK, etc.) do not need mid-invocation
+        # rescue — their corrective path fires at the next task
+        # boundary or via Level 3 CANCEL_REINVOKE.
+        if level is InterventionLevel.ABSORB and drift.kind in _ABSORB_NUDGE_KINDS:
+            nudge_msg = compose_corrective_user_message(
+                drift=drift,
+                refined_plan=session.plan,
+            )
+            session.pending_nudges.append(nudge_msg)
+            log.debug(
+                "DefaultSteerer._handle_drift: queued post-ABSORB nudge for kind=%s task=%s: %s",
+                drift.kind.value,
+                drift.current_task_id or "-",
+                nudge_msg,
+            )
+
+    # ------------------------------------------------------------------
+    # Level dispatch (#142) — nudge / steer / pause
+    # ------------------------------------------------------------------
+
+    async def _dispatch_nudge(self, drift: DriftEvent, session: Session) -> None:
+        """Level 2 dispatch: queue a soft follow-up message on the session.
+
+        The Runner's overlay loop (goldfive#141) picks up the queued
+        nudge at the next invocation boundary and sends it as a gentle
+        corrective user message. Until #141 lands, the queue is
+        observable but inert; nothing consumes it.
+        """
+        from goldfive.steerer import compose_corrective_user_message
+
+        msg = compose_corrective_user_message(
+            drift=drift,
+            refined_plan=session.plan,
+        )
+        session.pending_nudges.append(msg)
+        log.debug(
+            "DefaultSteerer: queued nudge for kind=%s task=%s: %s",
+            drift.kind.value,
+            drift.current_task_id or "-",
+            msg,
+        )
+
+    async def _dispatch_goldfive_steer_control(
+        self,
+        drift: DriftEvent,
+        session: Session,
+        *,
+        body_override: str = "",
+    ) -> bool:
+        """Mint and dispatch a ``GOLDFIVE_STEER`` ControlMessage.
+
+        Phase 2 of the path-duality fix. Replaces the dead
+        ``session.pending_corrective_message`` write at every
+        CANCEL_REINVOKE / promote-to-steer site so goldfive-authored
+        drift rides the same cancel-and-restart junction as
+        user-authored ``STEER``.
+
+        ``body_override``: optional text to use as the corrective
+        body. The promotion path passes its already-composed
+        :meth:`_compose_goldfive_steer_body` output; the Level 3
+        CANCEL_REINVOKE path leaves it empty and falls back to
+        :func:`compose_corrective_user_message` against the freshly
+        revised plan.
+
+        Returns ``True`` on successful dispatch, ``False`` on no
+        bound channel / send failure (best-effort — see
+        :meth:`DefaultSteerer._dispatch_goldfive_control`).
+        """
+        from goldfive.control import ControlKind, ControlMessage
+        from goldfive.steerer import compose_corrective_user_message
+
+        if body_override:
+            body = body_override
+        else:
+            body = compose_corrective_user_message(
+                drift=drift,
+                refined_plan=session.plan,
+            )
+        superseded_ids = (
+            [str(drift.current_task_id)] if drift.current_task_id else []
+        )
+        # Replacement task ids: pick the first PENDING task on the
+        # revised plan as the natural successor — the executor uses
+        # this to render an explicit "pick these up instead" block in
+        # the restart message.
+        replacement_ids: list[str] = []
+        plan = session.plan
+        if plan is not None:
+            for task in plan.tasks:
+                if task.status is TaskStatus.PENDING and task.id:
+                    replacement_ids.append(task.id)
+                    break
+        msg = ControlMessage(
+            kind=ControlKind.GOLDFIVE_STEER,
+            payload={
+                "drift_kind": drift.kind.value,
+                "drift_id": str(getattr(drift, "id", "") or ""),
+                "body": body,
+                "superseded_task_ids": superseded_ids,
+                "replacement_task_ids": replacement_ids,
+            },
+        )
+        # goldfive#254 — observation-only: skip the actual ControlMessage
+        # enqueue but log the would-be payload at INFO so operators can
+        # see what would have been dispatched (drift kind, task id, body).
+        # No cancel-and-restart fires on the executor; the live invocation
+        # continues against the prior plan.
+        if not self._steerer._should_inject():
+            _steerer_log.info(
+                "DefaultSteerer._dispatch_goldfive_steer_control: "
+                "observation_only=True — SKIPPING GOLDFIVE_STEER enqueue. "
+                "would_have_dispatched kind=%s task=%s drift_id=%s "
+                "superseded=%s replacement=%s body=%r",
+                drift.kind.value,
+                drift.current_task_id or "-",
+                str(getattr(drift, "id", "") or ""),
+                superseded_ids,
+                replacement_ids,
+                body[:200],
+            )
+            return False
+        landed = await self._steerer._dispatch_goldfive_control(msg)
+        log.debug(
+            "DefaultSteerer._dispatch_goldfive_steer_control: "
+            "kind=%s task=%s landed=%s",
+            drift.kind.value,
+            drift.current_task_id or "-",
+            landed,
+        )
+        return landed
+
+    async def _dispatch_goldfive_pause_control(
+        self,
+        drift: DriftEvent,
+        session: Session,
+        *,
+        reason: str,
+    ) -> bool:
+        """Mint and dispatch a ``GOLDFIVE_PAUSE_ESCALATE`` ControlMessage.
+
+        Phase 2 of the path-duality fix. Replaces the dead
+        ``session.paused_for_human_intervention = True`` flag-set at
+        every Level-4 / progress-stall / handler-exhausted escalation
+        site so the executor's pre-task loop blocks via the same
+        channel state as a user-issued ``PAUSE``.
+
+        Returns ``True`` on successful dispatch, ``False`` on no
+        bound channel / send failure.
+
+        goldfive#264 — observation-only carve-out. Under
+        ``SteeringConfig.observation_only`` the would-be
+        ``GOLDFIVE_PAUSE_ESCALATE`` is SKIPPED: dispatching it on the
+        channel sets ``goldfive_pause_message`` on the executor's
+        :class:`~goldfive.executors._control.ControlOutcome`, which in
+        turn drives ``_cancel_invoke_task`` and ends the overlay turn.
+        That kills the live invocation — exactly the enforcement
+        ``observation_only`` exists to suppress. The originating
+        ``HUMAN_INTERVENTION_REQUIRED`` drift emitted by the caller
+        (e.g. :meth:`_emit_handler_exhausted_escalation`,
+        :meth:`_emit_progress_stall_escalation`,
+        :meth:`_dispatch_pause_escalate`) is OUTSIDE this dispatch and
+        continues to fire — observers/sinks still see the escalation,
+        the operator can still react, but goldfive does NOT cancel the
+        in-flight invocation. Mirrors the gate pattern at
+        :meth:`_dispatch_goldfive_steer_control` (goldfive#254) and
+        :meth:`request_invocation_cancel`.
+
+        Live reproduction (2026-05-11, session
+        ``4538863f-0dea-4fe8-97b4-5f660ee2cb7f``): an OFF_TOPIC drift
+        under ``observation_only=True`` reached refine handler
+        exhaustion (#271 no-op-revision path), which called this
+        method, which dispatched the channel message, which cancelled
+        the in-flight invoke. The carve-out below stops that chain.
+        """
+        from goldfive.control import ControlKind, ControlMessage
+
+        if not self._steerer._should_inject():
+            _steerer_log.info(
+                "DefaultSteerer._dispatch_goldfive_pause_control: "
+                "observation_only=True — SKIPPING GOLDFIVE_PAUSE_ESCALATE "
+                "dispatch. would_have_dispatched kind=%s task=%s "
+                "drift_id=%s reason=%r",
+                drift.kind.value,
+                drift.current_task_id or "-",
+                str(getattr(drift, "id", "") or ""),
+                reason,
+            )
+            return False
+        msg = ControlMessage(
+            kind=ControlKind.GOLDFIVE_PAUSE_ESCALATE,
+            payload={
+                "reason": reason,
+                "drift_id": str(getattr(drift, "id", "") or ""),
+                "drift_kind": drift.kind.value,
+            },
+        )
+        landed = await self._steerer._dispatch_goldfive_control(msg)
+        log.debug(
+            "DefaultSteerer._dispatch_goldfive_pause_control: "
+            "kind=%s task=%s landed=%s reason=%r",
+            drift.kind.value,
+            drift.current_task_id or "-",
+            landed,
+            reason,
+        )
+        return landed
+
+    async def _dispatch_pause_escalate(
+        self,
+        drift: DriftEvent,
+        session: Session,
+    ) -> None:
+        """Level 4 dispatch: emit HUMAN_INTERVENTION_REQUIRED and pause.
+
+        Does NOT call ``planner.refine`` -- Level 4 signals that the
+        planner cannot recover. Phase 2 of the path-duality fix:
+        dispatches a ``GOLDFIVE_PAUSE_ESCALATE`` ControlMessage on the
+        bound channel so the executor's pre-task loop blocks via the
+        same channel state as a user ``PAUSE``. Pre-Phase-2 this
+        flipped ``session.paused_for_human_intervention = True`` — a
+        flag the executor read on its next iteration; the indirection
+        was synonymous with the channel signal but parallel-tracked
+        from the user-PAUSE path.
+
+        Emits a CRITICAL ``HUMAN_INTERVENTION_REQUIRED`` drift so
+        sinks / the UI can surface the pause and let the user decide
+        what to do.
+
+        When the drift reaching Level 4 is *already* a
+        ``HUMAN_INTERVENTION_REQUIRED`` (e.g. landed here via the
+        generic fallback), we pause but do not re-emit the same drift
+        a second time -- the original DriftDetected emission at the
+        top of :meth:`handle_drift` already carried the signal.
+        """
+        await self._dispatch_goldfive_pause_control(
+            drift,
+            session,
+            reason=(
+                f"pause_escalate from {drift.kind.value}: {drift.detail}"
+                if drift.detail
+                else f"pause_escalate from {drift.kind.value}"
+            ),
+        )
+        if drift.kind is DriftKind.HUMAN_INTERVENTION_REQUIRED:
+            # Already emitted at the top of _handle_drift; just pause.
+            return
+        # goldfive#271 PR1 — close the originating condition with
+        # ``human_intervention_required`` so consumers tracking the
+        # condition_id see the terminal lifecycle on the *original*
+        # condition, not just on the synthesized HUMAN_INTERVENTION
+        # row. The originating drift was already emitted at the top of
+        # ``_handle_drift`` (legacy path) under its own condition_id;
+        # this call swaps that condition's recorded lifecycle so a
+        # later get_active_drift returns the terminal state.
+        try:
+            origin_cid = _ostate.compute_condition_id(
+                kind=drift.kind,
+                task_id=str(getattr(drift, "current_task_id", "") or ""),
+                agent_id=str(getattr(drift, "current_agent_id", "") or ""),
+                turn_id=str(getattr(session, "run_id", "") or ""),
+            )
+            _ostate.escalate_drift_to_human_intervention(session.state, origin_cid)
+        except Exception as exc:  # noqa: BLE001
+            log.debug("DefaultSteerer: drift-lifecycle escalate skipped (%s)", exc)
+        escalation = DriftEvent(
+            kind=DriftKind.HUMAN_INTERVENTION_REQUIRED,
+            severity=DriftSeverity.CRITICAL,
+            detail=(
+                f"escalated from {drift.kind.value}: {drift.detail}"
+                if drift.detail
+                else f"escalated from {drift.kind.value}"
+            ),
+            current_task_id=drift.current_task_id,
+            current_agent_id=drift.current_agent_id,
+        )
+        # Emit directly; do NOT go back through _handle_drift (would
+        # infinite-loop at CRITICAL).
+        await self._emit_drift_detected(session, escalation)
+
+    # ------------------------------------------------------------------
+    # Adapter cancel-reason tagging
+    # ------------------------------------------------------------------
+    #
+    # Symbolic cancel-reason tags — mirror
+    # :mod:`goldfive.adapters.adk` constants but duplicated as plain
+    # strings here to avoid a hard import of the optional ADK adapter
+    # module from the provider-agnostic steerer. Keep in sync with
+    # :data:`goldfive.adapters.adk.SYMBOLIC_REASON_USER_STEER` etc.
+    _ADAPTER_CANCEL_REASON_USER_STEER: str = "user_steer"
+
+    def _tag_adapter_cancel_reason(
+        self, drift: DriftEvent, *, session: Session | None = None
+    ) -> None:
+        """Set the next adapter cancel reason based on ``drift.kind``.
+
+        USER_STEER drift -> ``"user_steer"``. Other kinds currently leave
+        the tag unset so the adapter falls through to the generic
+        content variant. Tolerates adapters that don't carry the
+        attribute (no-op) and an unbound adapter (no-op). See
+        goldfive#139.
+
+        Routes the write through
+        :meth:`ADKAdapter.set_next_cancel_reason` when the adapter
+        exposes that helper (PR #294 audit / goldfive#271 follow-up)
+        so the tag is keyed by ``session.id`` and cannot bleed across
+        concurrent goldfive sessions sharing one adapter. Falls back
+        to the bare attribute write for adapters / stubs that predate
+        the helper.
+
+        The goldfive-steer-unification promotion path uses a separate
+        helper (:meth:`_tag_adapter_cancel_reason_for_promotion`) to
+        stamp a ``"goldfive_<drift_kind>"`` reason when promoting a
+        detector drift to a full steer; keeping the two call sites
+        distinct avoids muddling the pre-unification tag semantics for
+        unpromoted paths.
+        """
+        adapter = self._steerer._adapter
+        if adapter is None:
+            return
+        if drift.kind is DriftKind.USER_STEER:
+            reason = self._ADAPTER_CANCEL_REASON_USER_STEER
+        else:
+            return
+        self._write_adapter_cancel_reason(adapter, reason, session)
+
+    def _tag_adapter_cancel_reason_for_promotion(
+        self, drift: DriftEvent, *, session: Session | None = None
+    ) -> str:
+        """Stamp a goldfive-specific cancel reason on the bound adapter.
+
+        Returns the reason string stamped (or synthesised) so callers
+        can record it on the session for downstream observability.
+        Mirrors :meth:`_tag_adapter_cancel_reason` semantics: adapters
+        without the per-session helper are tolerated.
+        """
+        reason = f"goldfive_{drift.kind.name.lower()}"
+        adapter = self._steerer._adapter
+        if adapter is None:
+            return reason
+        self._write_adapter_cancel_reason(adapter, reason, session)
+        return reason
+
+    @staticmethod
+    def _write_adapter_cancel_reason(adapter: Any, reason: str, session: Session | None) -> None:
+        """Route the cancel-reason tag through the session-aware helper.
+
+        Falls back to the legacy bare-attribute write for adapters /
+        stubs that don't expose :meth:`set_next_cancel_reason`. See
+        :meth:`ADKAdapter.set_next_cancel_reason` for the rationale.
+        """
+        setter = getattr(adapter, "set_next_cancel_reason", None)
+        if callable(setter) and session is not None:
+            try:
+                setter(session, reason)
+                return
+            except Exception as exc:  # noqa: BLE001
+                log.debug("DefaultSteerer: set_next_cancel_reason raised: %s", exc)
+        try:
+            adapter._next_cancel_reason = reason
+        except Exception as exc:  # noqa: BLE001
+            log.debug(
+                "DefaultSteerer: could not tag adapter cancel reason: %s",
+                exc,
+            )
+
+    async def _request_adapter_cancel(self, reason: str) -> None:
+        """Invoke the optional ``adapter.request_cancel(reason)`` hook.
+
+        goldfive#241 — a goldfive-promoted steer needs the in-flight
+        LLM call to stop NOW so the contaminated reasoning / tool
+        calls don't keep writing to the session while we queue the
+        restart. The ADK adapter exposes :meth:`ADKAdapter.request_cancel`
+        which fires ``task.cancel()`` on the asyncio task driving
+        ``runner.run_async`` so the stream raises ``CancelledError``
+        and the adapter's standard heal path runs with the already-
+        stamped ``_next_cancel_reason`` tag.
+
+        Optional protocol: adapters that don't implement the method
+        (Claude adapter, callable adapter, test stubs without live
+        invocations) keep the legacy deferred-cancel semantics —
+        ``_next_cancel_reason`` is still tagged, the restart message
+        is still queued, and the next executor checkpoint still
+        terminates the invocation. Tolerates an unbound adapter and
+        swallows every failure so a best-effort cancel cannot break
+        the promotion path.
+        """
+        adapter = self._steerer._adapter
+        if adapter is None:
+            return
+        fn = getattr(adapter, "request_cancel", None)
+        if not callable(fn):
+            return
+        try:
+            result = fn(reason)
+            if inspect.isawaitable(result):
+                await result
+        except Exception as exc:  # noqa: BLE001
+            log.debug(
+                "DefaultSteerer._request_adapter_cancel(reason=%r): adapter raised: %s",
+                reason,
+                exc,
+            )
+
+    # ------------------------------------------------------------------
+    # Cooperative cancellation (goldfive#251 Stream C / 7a)
+    # ------------------------------------------------------------------
+
+    def _is_late_drift_for_terminated_invocation(
+        self, drift: DriftEvent, session: Session
+    ) -> bool:
+        """Return True iff a goldfive-authored drift's target is gone (goldfive#319).
+
+        Background reasoning-judge tasks (goldfive#251) run off the
+        critical path so the adapter's model-response callback can return
+        before the LLM judge finishes. With goldfive#319's removal of the
+        per-turn cancel-drain, a slow judge spawned in turn N may now
+        produce its verdict in turn N+1 — well after the original agent
+        invocation has terminated. Routing such a verdict through the
+        cancel + ladder dispatch is a category error: it could cancel an
+        unrelated invocation or trigger a refine against a plan whose
+        offending step is already complete. The drift is still emitted
+        on the sink (observability preserved); this guard short-circuits
+        the dispatch.
+
+        The check uses :class:`StateStore` as the live registry
+        of in-flight invocations (Phase 3.5 component 1, goldfive#271).
+        Two conditions count as "late":
+
+        * **No active invocations** — every agent has finished its turn
+          and any drift currently being handled is by definition stale.
+        * **Cancel-pending on the session** (goldfive#242) — a previous
+          drift already requested a cooperative cancel for one or more
+          invocations. The active-task registry takes 4-8s to drain
+          while ADK winds those invocations down; during that window
+          any newly-arriving goldfive-authored drift would dispatch a
+          refine against an effectively-dead session. Stamping the
+          cancel-pending flag synchronously at
+          :meth:`request_invocation_cancel` time closes that race.
+
+        User-authored drifts (USER_STEER / USER_CANCEL / USER_PAUSE)
+        always bypass this guard — they are forward-looking operator
+        directives, not tied to a specific in-flight invocation.
+        """
+        # User-authored drifts always pass through. ``authored_by`` was
+        # normalised at the top of :meth:`handle_drift`.
+        if (drift.authored_by or "").lower() == "user":
+            return False
+        try:
+            from goldfive.state_store import StateStore
+
+            store = StateStore.for_session(session)
+            active = store.active_invocation_ids()
+            cancel_pending = store.cancel_requested_invocation_ids()
+        except Exception as exc:  # noqa: BLE001 — defensive
+            log.debug(
+                "DefaultSteerer._is_late_drift_for_terminated_invocation: "
+                "active_invocation_ids lookup raised (treating as not-late): %s",
+                exc,
+            )
+            return False
+        # Symmetric predicate: late when the active list is empty OR
+        # any cancel is pending on the session. The cancel-pending
+        # branch closes the iter-11D race (goldfive#242) where the
+        # cancel-request has landed but ADK hasn't yet finished
+        # winding down the cancelled invocation.
+        return (not active) or bool(cancel_pending)
+
+    def _resolve_active_invocation_ids(self, drift: DriftEvent, session: Session) -> list[str]:
+        """Resolve which invocation_id(s) a cancel should target.
+
+        Returns an ordered list of invocation ids that are "active"
+        with respect to the triggering drift. The primary source is
+        the reconciler's invocation bookkeeping (goldfive#151
+        introduced the ``_invocation_agent`` / ``_invocation_parent``
+        maps). When the reconciler is unavailable or empty, falls
+        back to the drift's ``current_agent_id``-keyed invocation
+        (best effort via the adapter's active-context invocation id)
+        and finally returns an empty list.
+
+        Tree-agnostic: the method does NOT special-case "the
+        coordinator" or "the root agent" — it targets whichever
+        invocation matches the drift's context and lets the plugin's
+        child-propagation logic flag the rest of the sub-tree.
+        """
+        candidates: list[str] = []
+        reconciler = getattr(session, "_reconciler", None)
+        if reconciler is None:
+            # The steerer doesn't hold a direct reference to the
+            # reconciler; the adapter's plugin does. Walk it via the
+            # adapter when the plugin exposes the attribute.
+            adapter = self._steerer._adapter
+            plugin = getattr(adapter, "_plugin", None) if adapter is not None else None
+            reconciler = getattr(plugin, "_reconciler", None) if plugin is not None else None
+        if reconciler is not None:
+            try:
+                inv_agent = getattr(reconciler, "_invocation_agent", None)
+                if isinstance(inv_agent, Mapping) and drift.current_agent_id:
+                    # Match by agent name — most drifts carry
+                    # ``current_agent_id`` set to the running agent's name.
+                    for inv_id, agent_name in inv_agent.items():
+                        if agent_name == drift.current_agent_id and inv_id:
+                            candidates.append(str(inv_id))
+            except Exception as exc:  # noqa: BLE001
+                log.debug(
+                    "DefaultSteerer._resolve_active_invocation_ids: reconciler lookup raised: %s",
+                    exc,
+                )
+        # Fallback: the adapter's plugin pins a top-level invocation_id
+        # for the currently-driving dispatch. When the reconciler lookup
+        # produced nothing, the top-level id is the best we can do —
+        # cancel propagation from there will flag any sub-invocations.
+        if not candidates:
+            adapter = self._steerer._adapter
+            plugin = getattr(adapter, "_plugin", None) if adapter is not None else None
+            top = str(getattr(plugin, "_top_invocation_id", "") or "")
+            if top:
+                candidates.append(top)
+        return candidates
+
+    async def request_invocation_cancel(
+        self,
+        *,
+        drift: DriftEvent,
+        session: Session,
+        cancel_inflight_task: bool = False,
+    ) -> list[str]:
+        """Flag the invocation(s) associated with ``drift`` for
+        cooperative cancellation (goldfive#251 Stream C / 7a).
+
+        Called from :meth:`handle_drift` and
+        :meth:`_promote_drift_to_steer` when the drift's severity is
+        CRITICAL — the only tier on the ladder that reaches the hard
+        cancel per the severity decision. INFO / WARNING drifts flow
+        through their usual nudge / absorb paths without touching
+        this method.
+
+        Writes a :class:`~goldfive.types.CancellationRequest` onto the
+        adapter's plugin state for every resolved active invocation
+        id. The plugin propagates to children automatically. Returns
+        the list of flagged invocation ids (including children) for
+        observability; callers can log / sink-emit from the list.
+
+        When ``cancel_inflight_task=True`` (goldfive#271 follow-up —
+        v15 concurrent-invocation bug), the plugin ALSO fires
+        ``task.cancel()`` on the registered asyncio.Task driving each
+        flagged invocation, deferred via ``loop.call_soon`` so an
+        inline same-task caller still completes its current emission
+        work before the cancel lands. Default False so the existing
+        pre-refine cancel paths keep their flag-only semantics; the
+        post-refine helper :meth:`_cancel_inflight_for_revision`
+        opts in explicitly so the cancel only fires AFTER a
+        superseding plan has been installed.
+
+        Guard rails:
+
+        * No-op when no adapter is bound.
+        * No-op when no active invocation can be resolved (e.g. the
+          drift was synthesized before any agent turn started) — this
+          is the "empty invocation-id guard" called out in the brief.
+        * Tolerates missing plugin methods (third-party adapters that
+          don't implement :meth:`request_invocation_cancel`) by
+          falling through silently; the rest of the ladder (refine,
+          restart message) still runs and eventually catches up at
+          the next task boundary.
+        * Plugins whose ``request_invocation_cancel`` predates
+          ``cancel_inflight_task`` (TypeError on the kwarg) fall back
+          to the kwarg-less call so older third-party plugins don't
+          break — the task-cancel step is silently skipped.
+
+        Observation-only mode (goldfive#254): when
+        :meth:`DefaultSteerer._should_inject` is ``False`` this method
+        returns ``[]`` without consulting the plugin or stamping
+        ``cancel_requested_invocation_ids``. Logged at INFO so an
+        operator can see WHAT would have been cancelled (drift kind,
+        task / agent id) without the cancel actually firing on the
+        live invocation.
+        """
+        if not self._steerer._should_inject():
+            _steerer_log.info(
+                "DefaultSteerer.request_invocation_cancel: "
+                "observation_only=True — SKIPPING cancel for "
+                "drift kind=%s severity=%s agent=%s task=%s",
+                drift.kind.value,
+                drift.severity.value,
+                drift.current_agent_id or "-",
+                drift.current_task_id or "-",
+            )
+            return []
+        adapter = self._steerer._adapter
+        if adapter is None:
+            return []
+        plugin = getattr(adapter, "_plugin", None)
+        if plugin is None:
+            return []
+        fn = getattr(plugin, "request_invocation_cancel", None)
+        if not callable(fn):
+            return []
+        invocation_ids = self._resolve_active_invocation_ids(drift, session)
+        # Stamp the cancel-pending flag SYNCHRONOUSLY before any
+        # plugin / async work (goldfive#242). The active-task
+        # registry takes 4-8s to drain while ADK winds the cancelled
+        # invocations down; during that window the late-drift gate
+        # would otherwise see ``active_invocation_ids()`` non-empty
+        # and let a freshly-arriving goldfive-authored drift dispatch
+        # a refine against an effectively-dead session. Flipping the
+        # flag here closes the race: any drift handled after this
+        # point sees ``cancel_requested_invocation_ids()`` non-empty
+        # via :meth:`_is_late_drift_for_terminated_invocation` and
+        # short-circuits.
+        if invocation_ids:
+            try:
+                from goldfive.state_store import StateStore
+
+                store = StateStore.for_session(session)
+                for inv_id in invocation_ids:
+                    store.mark_invocation_cancel_requested(inv_id)
+            except Exception as exc:  # noqa: BLE001 — defensive
+                log.debug(
+                    "DefaultSteerer.request_invocation_cancel: "
+                    "cancel-pending stamp raised (continuing): %s",
+                    exc,
+                )
+        if not invocation_ids:
+            # Empty invocation-id guard — drift has no identifiable
+            # in-flight invocation. Don't fabricate one; the cancel
+            # would misfire on whatever invocation happens to share a
+            # blank id. The drift still observed, refine still runs;
+            # cancel is a best-effort add-on.
+            log.debug(
+                "DefaultSteerer.request_invocation_cancel: no active invocation "
+                "for drift kind=%s agent=%s task=%s — skipping cancel",
+                drift.kind.value,
+                drift.current_agent_id or "-",
+                drift.current_task_id or "-",
+            )
+            return []
+        # Build the request once and reuse for every targeted id so
+        # sink events from propagation share a common fingerprint.
+        import time as _time_mod
+
+        request = CancellationRequest(
+            invocation_id=invocation_ids[0],
+            reason=self._cancel_reason_for_drift(drift),
+            severity=drift.severity,
+            drift_id=str(getattr(drift, "id", "") or ""),
+            drift_kind=drift.kind.value,
+            requested_at_ms=int(_time_mod.time() * 1000),
+            detail=(drift.detail or "")[:200],
+        )
+        flagged: list[str] = []
+        for inv_id in invocation_ids:
+            try:
+                result = fn(
+                    invocation_id=inv_id,
+                    request=request,
+                    cancel_inflight_task=cancel_inflight_task,
+                )
+            except TypeError:
+                # Older plugin without the ``cancel_inflight_task``
+                # kwarg (third-party / pre-#271-follow-up). Fall back
+                # to the legacy signature; the task-cancel step is
+                # silently skipped, but the flag-only contract is
+                # preserved.
+                try:
+                    result = fn(invocation_id=inv_id, request=request)
+                except Exception as exc:  # noqa: BLE001
+                    log.debug(
+                        "DefaultSteerer.request_invocation_cancel: "
+                        "plugin.request_invocation_cancel(%s) raised: %s",
+                        inv_id,
+                        exc,
+                    )
+                    continue
+            except Exception as exc:  # noqa: BLE001
+                log.debug(
+                    "DefaultSteerer.request_invocation_cancel: "
+                    "plugin.request_invocation_cancel(%s) raised: %s",
+                    inv_id,
+                    exc,
+                )
+                continue
+            if isinstance(result, list):
+                flagged.extend(str(x) for x in result)
+            else:
+                flagged.append(inv_id)
+        if flagged:
+            _steerer_log.info(
+                "DefaultSteerer.request_invocation_cancel: flagged "
+                "invocations=%s for drift kind=%s severity=%s",
+                flagged,
+                drift.kind.value,
+                drift.severity.value,
+            )
+        return flagged
+
+    @staticmethod
+    def _should_request_cancel_for_drift(drift: DriftEvent) -> bool:
+        """Decide whether a drift warrants a cooperative cancel.
+
+        Severity ladder (goldfive#251 design decision):
+
+        * ``DriftSeverity.INFO`` — never cancels. Info drifts are
+          either periodic-check signals or soft one-shots; cancel
+          would be disproportionate.
+        * ``DriftSeverity.WARNING`` — never cancels. Warning drifts
+          route to the existing ABSORB / NUDGE ladder paths; the
+          refined plan lands on the next task boundary without
+          preempting the in-flight turn.
+        * ``DriftSeverity.CRITICAL`` — cancels. The in-flight turn's
+          output is likely to contaminate its parent's transcript
+          (stale prompt, wrong scope, broken tool); short-circuit
+          cleanly and let the parent see ``{"status": "cancelled"}``.
+
+        User-authored drifts (``USER_STEER`` / ``USER_CANCEL`` /
+        ``USER_PAUSE``) bypass the severity gate — an operator
+        directive must be honoured even when the ControlMessage-to-
+        DriftEvent coercion landed on a lower severity tier.
+        """
+        if drift.kind in (
+            DriftKind.USER_STEER,
+            DriftKind.USER_CANCEL,
+            DriftKind.USER_PAUSE,
+        ):
+            return True
+        return drift.severity is DriftSeverity.CRITICAL
+
+    @staticmethod
+    def _cancel_reason_for_drift(drift: DriftEvent) -> str:
+        """Map a drift into a short symbolic reason for the
+        :class:`~goldfive.types.CancellationRequest`.
+
+        USER_STEER / USER_CANCEL / USER_PAUSE get the matching
+        ``"user_*"`` shorthand; everything else uses ``"drift"`` as
+        the generic tag. The reason is OPERATOR-visible only (lives
+        on the InvocationCancelled sink event), so this string is
+        free to be descriptive without prompt-injection concerns.
+        """
+        kind = drift.kind
+        if kind is DriftKind.USER_STEER:
+            return "user_steer"
+        if kind is DriftKind.USER_CANCEL:
+            return "user_cancel"
+        if kind is DriftKind.USER_PAUSE:
+            return "user_pause"
+        return "drift"
+
+    async def _cancel_inflight_for_revision(self, drift: DriftEvent, session: Session) -> list[str]:
+        """Cancel the in-flight invocation(s) that produced ``drift``.
+
+        Called from every drift-driven PlanRevised emission path right
+        after the revised plan has been applied to the session and
+        BEFORE the ``PlanRevised`` event is emitted. Closes the gap
+        behind the v15 concurrent-invocation bug: a ``refine_steer``
+        call (10+ minutes on a slow planner) used to overlap the
+        coordinator's invocation for its full duration because the
+        existing cancel-state flag only gates SUBSEQUENT callbacks —
+        the already-running LLM streaming call kept generating output
+        that triggered more drift, looping the refine.
+
+        After Option A (goldfive#271 follow-up), turn-1 first-plan
+        installs no longer reach this path:
+        :meth:`install_initial_plan` skips it directly because there
+        is no in-flight invocation to cancel on a fresh session. Every
+        drift-driven install (refine from drift, refine_steer from
+        goldfive-steer promotion, operator USER_STEER from a real
+        ControlMessage, NEW_WORK_DISCOVERED from an N+1 user message)
+        flows through this method. The plugin's
+        :meth:`request_invocation_cancel` then writes the cancel-state
+        flag (sticky-gate from PR #299) AND fires ``task.cancel()`` on
+        the registered asyncio.Task (goldfive#271 follow-up) so the
+        coordinator's in-flight LLM call observes ``CancelledError``
+        within ~one event-loop tick instead of ~the LLM-call's
+        full duration.
+
+        Supersede contract (Bug A fix from v22 validation): every call
+        to this method represents a goldfive-INTERNAL cancel — the
+        revised plan has just been applied, and the cancel is the
+        mechanism by which the in-flight agent is switched onto it.
+        Stamps ``session._supersede_pending = True`` BEFORE initiating
+        the cancel so the executor's overlay loop
+        (:meth:`SequentialExecutor._run_overlay` cancelled branch) can
+        distinguish this internal supersede from an external cancel
+        (USER_CANCEL via control channel, asyncio.CancelledError from
+        above) and restart the passthrough loop with the new plan
+        instead of aborting the turn. The executor consumes and clears
+        the flag; an unconsumed flag (e.g. cancel never lands because
+        the invocation already completed) is harmless — the next
+        cancel-branch entry will see and clear it, or the run finishes
+        normally and the Session is discarded.
+
+        Best-effort: an unbound adapter, a non-ADK adapter without
+        :meth:`request_invocation_cancel`, or an empty resolved
+        invocation-id list each result in a no-op (the refined plan
+        still lands; the in-flight invocation simply runs to
+        completion under the older, less aggressive contract). The
+        supersede flag is still stamped in the no-op case — it costs
+        nothing and a downstream overlay that DOES observe a cancel
+        from a separate path stays correctly classified.
+        """
+        # Stamp the supersede marker so the overlay loop can
+        # distinguish this internal cancel from an external one. See
+        # the supersede-contract paragraph in the docstring above.
+        try:
+            session._supersede_pending = True  # type: ignore[attr-defined]
+        except Exception as exc:  # noqa: BLE001 — flag is best-effort
+            log.debug(
+                "DefaultSteerer._cancel_inflight_for_revision: "
+                "could not stamp supersede flag on session: %s",
+                exc,
+            )
+        try:
+            return await self.request_invocation_cancel(
+                drift=drift,
+                session=session,
+                cancel_inflight_task=True,
+            )
+        except Exception as exc:  # noqa: BLE001 — cancel is best-effort
+            log.debug(
+                "DefaultSteerer._cancel_inflight_for_revision: "
+                "request_invocation_cancel raised: %s",
+                exc,
+            )
+            return []
+
+    # ------------------------------------------------------------------
+    # goldfive-steer-unification: promotion policy + handler
+    # ------------------------------------------------------------------
+
+    # Drift kinds eligible for ladder-promoted steer treatment when
+    # goldfive-authored. Mirrors the "content drifts the coordinator
+    # acknowledges but doesn't correct" list from the unification
+    # design brief: off-topic / intent / unexpected output /
+    # confabulation / loop kinds. Other detector kinds (SCHEMA_VIOLATION,
+    # REFINE_VALIDATION_FAILED, REPEATED_FAILURE, GOAL_DRIFT, …) keep
+    # their pre-unification ladder mapping so escalation / repeated-
+    # failure semantics aren't rerouted into the cancel-in-flight path.
+    _GOLDFIVE_STEER_ELIGIBLE_KINDS: frozenset[DriftKind] = frozenset(
+        {
+            DriftKind.OFF_TOPIC,
+            DriftKind.INTENT_DIVERGENCE,
+            DriftKind.UNEXPECTED_OUTPUT,
+            DriftKind.CONFABULATION_RISK,
+            DriftKind.LOOPING_REASONING,
+            DriftKind.LOOPING_TOOL_CALL,
+            DriftKind.PLAN_DIVERGENCE,
+        }
+    )
+
+    def _severity_meets_promotion_threshold(self, severity: DriftSeverity) -> bool:
+        """True iff ``severity`` satisfies the configured promotion threshold."""
+        threshold = self._steerer._goldfive_steer_threshold
+        if threshold == "off":
+            return False
+        if threshold == "critical":
+            return severity is DriftSeverity.CRITICAL
+        # "warning" — promote WARNING and CRITICAL.
+        return severity in (DriftSeverity.WARNING, DriftSeverity.CRITICAL)
+
+    def _should_promote_to_steer(self, drift: DriftEvent, session: Session) -> bool:
+        """Evaluate the drift against the unification promotion policy.
+
+        Returns ``True`` iff the drift should be dispatched through
+        :meth:`_promote_drift_to_steer` instead of the legacy passive
+        ladder. Side-effect: stamps ``drift.suppressed_by_user_steer``
+        when a fresh user steer is blocking promotion so the subsequent
+        ``DriftDetected`` emission reflects the suppression decision.
+
+        The policy:
+
+        1. User-authored drifts (USER_STEER / USER_CANCEL / USER_PAUSE)
+           keep their pre-unification handling — USER_STEER already
+           routes through the refine path with cancel-in-flight wired
+           by the executor. Return ``False``.
+        2. The drift kind must be in
+           :data:`_GOLDFIVE_STEER_ELIGIBLE_KINDS` — other kinds keep
+           their legacy ladder mapping.
+        3. The severity must clear the configured ``threshold``.
+        4. If a user-authored steer is within the freshness window
+           (``suppression_window_turns`` turns), stamp the suppression
+           flag and return ``False``. Otherwise return ``True``.
+        """
+        if drift.kind in self._USER_AUTHORED_DRIFT_KINDS:
+            return False
+        authored_by = self._resolve_authored_by(drift)
+        if authored_by != "goldfive":
+            return False
+        if drift.kind not in self._GOLDFIVE_STEER_ELIGIBLE_KINDS:
+            return False
+        if not self._severity_meets_promotion_threshold(drift.severity):
+            return False
+        # Consult the active user steer freshness window.
+        # Phase 1 of goldfive#271 — read through StateStore so
+        # the active-steer slot reads from a single named accessor; the
+        # underlying ``_ostate.read`` calls still funnel through the
+        # goldfive Session.state dict, just behind a typed surface.
+        window = self._steerer._goldfive_steer_suppression_window_turns
+        if window > 0:
+            from goldfive.state_store import StateStore
+
+            active = StateStore.for_session(session).get_active_steer()
+            if active is not None and active.source.lower() == "user":
+                current_turn = int(getattr(session, "_next_sequence", 0) or 0)
+                age = current_turn - active.at_turn
+                if 0 <= age < window:
+                    drift.suppressed_by_user_steer = True
+                    _steerer_log.info(
+                        "goldfive steer suppressed: user steer %r is active "
+                        "(age=%d turns, window=%d)",
+                        active.body,
+                        age,
+                        window,
+                    )
+                    return False
+        return True
+
+    async def _promote_drift_to_steer(self, drift: DriftEvent, session: Session) -> None:
+        """Promote a goldfive-detected drift into a full steer.
+
+        Ordered side effects (mirrors the USER_STEER path):
+
+        1. Tag the bound adapter's ``_next_cancel_reason`` with a
+           ``"goldfive_<drift_kind>"`` symbolic reason so the in-flight
+           invocation's synthetic ``function_response`` carries an
+           LLM-actionable explanation.
+        2. Stamp ``goldfive.active_steer.*`` onto ``session.state``
+           (body = derived from :meth:`_compose_goldfive_steer_body`,
+           author = ``"goldfive"``, source = ``"goldfive"``).
+        3. Record ``drift.id`` in ``goldfive.processed_steer_ids`` so
+           the same drift cannot re-promote on a delivery retry.
+        4. Call :meth:`LLMPlanner.refine_steer` (or the generic
+           ``planner.refine`` fallback when the planner doesn't expose
+           the goldfive-specific entry point) with ``source="goldfive"``
+           semantics so the refine prompt frames the pivot as a
+           correction, not as an operator directive.
+        5. Install the revised plan + emit ``PlanRevised``.
+
+        Note on cancel-in-flight: the actual ``task.cancel()`` on the
+        adapter invocation is the executor's responsibility
+        (:meth:`SequentialExecutor._invoke_with_control` performs it
+        when a ``STEER`` ControlMessage arrives). The steerer tags the
+        adapter and queues a restart message so that the **next** time
+        the executor reaches a cancel / steer checkpoint (either
+        because a sink callback requested cancel, or because the
+        overlay loop picks up the pending restart message), the
+        contaminated invocation is preempted. For the common case
+        where the drift is detected from a mid-invocation reasoning
+        block and the overlay loop is already streaming, the queued
+        restart message reaches the LLM on the next turn — cancel
+        semantics identical to USER_STEER.
+
+        Audit issue #402 (preserved, not fixed): the dispatch order
+        below sends the ``GOLDFIVE_STEER`` ControlMessage BEFORE the
+        refine + plan install, so the dispatch payload carries the
+        prior plan's task ids in ``superseded_task_ids`` /
+        ``replacement_task_ids``. The fix for that ordering lives in a
+        separate queued PR; the refactor must not alter the ordering
+        or the bug surface changes between PRs.
+        """
+        from goldfive.steerer import (
+            RefineExhausted,
+            _planner_refine_accepts_available_agents,
+        )
+
+        # 1. Tag adapter cancel reason.
+        cancel_reason = self._tag_adapter_cancel_reason_for_promotion(drift, session=session)
+        # Session-visible cancel prefix so ``_mark_cancelled_if_live``
+        # stamps it on any TaskCancelled the executor emits for the
+        # in-flight task as part of the promotion.
+        try:
+            session._last_cancel_reason_prefix = cancel_reason
+        except Exception:  # noqa: BLE001
+            pass
+        # 1a. NOTE (#241 emergency revert): previously we fired
+        # ``adapter.request_cancel(reason)`` here to terminate the
+        # in-flight LLM call immediately. In practice that
+        # ``task.cancel()`` propagated a ``CancelledError`` past the
+        # executor's invocation-scope catch and killed the entire run
+        # — observed as ``run_aborted`` immediately after a
+        # goldfive-detected drift. Reverted to the pre-#241 deferred-
+        # cancel semantics: we stamp ``_next_cancel_reason`` (above)
+        # and queue a restart message; the executor loop sees the
+        # queue at the next invocation boundary and resumes with the
+        # refined plan. The taint of letting the in-flight call run
+        # to completion is a lesser evil than aborting the run.
+        # Proper fix (future): scope the cancel to the LLM stream
+        # only, or catch ``CancelledError`` at the goldfive-steer
+        # boundary and continue.
+        # 2. Stamp active-steer state + compose the restart body.
+        at_turn = int(getattr(session, "_next_sequence", 0) or 0)
+        body = self._compose_goldfive_steer_body(drift)
+        try:
+            _ostate.set_active_steer(
+                session.state,
+                body=body,
+                at_turn=at_turn,
+                author="goldfive",
+                source="goldfive",
+            )
+        except Exception as exc:  # noqa: BLE001
+            log.debug(
+                "DefaultSteerer._promote_drift_to_steer: set_active_steer raised: %s",
+                exc,
+            )
+        # Phase 2 of the path-duality fix: dispatch a
+        # ``GOLDFIVE_STEER`` ControlMessage on the bound channel so
+        # the executor's invoke loop cancels the in-flight invocation
+        # and restarts the passthrough with a ``[GOLDFIVE STEERING
+        # CONTROL …]`` framed corrective. The body, drift kind, and
+        # superseded task ids ride the message payload; the executor
+        # composes the restart text from those fields. Pre-Phase-2
+        # this wrote ``session.pending_corrective_message`` — a
+        # write-only slot that left the coordinator blind to the plan
+        # swap.
+        #
+        # AUDIT #402: this dispatch happens BEFORE the refine + plan
+        # install below — the payload carries the prior plan's task
+        # ids. Preserved as-is.
+        await self._dispatch_goldfive_steer_control(
+            drift, session, body_override=body
+        )
+        # 3. Record the drift id in processed_steer_ids so a redelivery
+        # (same drift id) doesn't re-cancel / re-refine.
+        drift_id = str(getattr(drift, "id", "") or "")
+        if drift_id:
+            try:
+                _ostate.record_processed_steer_id(session.state, drift_id)
+            except Exception as exc:  # noqa: BLE001
+                log.debug(
+                    "DefaultSteerer._promote_drift_to_steer: record_processed_steer_id raised: %s",
+                    exc,
+                )
+        # 4. Route to planner.refine_steer (source="goldfive") — falls
+        # back to planner.refine for planners that don't expose the
+        # goldfive-specific entry point.
+        if self._steerer._planner is None or session.plan is None:
+            return
+        # Outcome-based gate (goldfive#215 iter-8 P2). Mirror of the
+        # gate in ``_handle_drift``: skip refine_steer when (kind, task)
+        # already has a terminal outcome on this turn. USER_STEER /
+        # USER_CANCEL / GOAL_DRIFT bypass.
+        if drift.kind not in self._USER_OR_TRAJECTORY_DRIFT_KINDS:
+            outcome_key = (drift.kind.value, drift.current_task_id or "")
+            outcome = session.refine_outcomes.get(outcome_key)
+            if outcome is not None:
+                if outcome.state == "succeeded":
+                    log.debug(
+                        "refine_steer skipped: prior succeeded outcome (kind=%s task=%r)",
+                        drift.kind.value,
+                        drift.current_task_id,
+                    )
+                    return
+                if outcome.fail_count >= self._steerer.REFINE_FAILURE_THRESHOLD:
+                    log.debug(
+                        "refine_steer skipped: failure threshold reached "
+                        "(kind=%s task=%r count=%d)",
+                        drift.kind.value,
+                        drift.current_task_id,
+                        outcome.fail_count,
+                    )
+                    return
+        # Progress-based escalation (goldfive#271). Orthogonal to the
+        # outcome gate: see parallel check in ``_handle_drift``.
+        if self._is_task_progress_stalled(drift, session):
+            await self._emit_progress_stalled_escalation(drift, session)
+            return
+        # ContextVar plumbing for the planner-side drift-emitter and
+        # span-context callbacks; per-async-task so concurrent runs
+        # sharing this Steerer keep their session pointers isolated.
+        _active_session_token = self._steerer._active_session_var.set(session)
+        # goldfive a4: same attempt-id correlation contract as
+        # ``_handle_drift``.
+        attempt_id = self._steerer._new_attempt_id()
+        await self._steerer._emit_refine_attempted(session, drift, attempt_id=attempt_id)
+        # Resolve the registry constraint (goldfive#151) the same way
+        # ``_handle_drift`` does so the goldfive steer refine honours it.
+        planner = self._steerer._planner
+        available_agents: Any = None
+        adapter = self._steerer._adapter
+        if adapter is not None:
+            tree = getattr(adapter, "available_agents_tree", None)
+            if isinstance(tree, list) and tree:
+                available_agents = list(tree)
+            else:
+                flat = getattr(adapter, "available_agents", None)
+                if flat:
+                    available_agents = list(flat)
+        # Phase 3.5 (goldfive#271) tripwire wrapper — see §C4.
+        with _state_audit.cancellation_stash_audited(
+            "DefaultSteerer._promote_drift_to_steer.refine"
+        ):
+            try:
+                # Call ``planner.refine_steer`` when available; fall back
+                # to ``planner.refine``. The fallback exists for test
+                # stubs / third-party planners that don't expose the
+                # goldfive-specific entry point — the generic path is
+                # better than no refine at all.
+                refine_steer = getattr(planner, "refine_steer", None)
+                if callable(refine_steer):
+                    revised = await refine_steer(
+                        plan=session.plan,
+                        drift=drift,
+                        goals=list(session.goals),
+                        available_agents=available_agents,
+                    )
+                elif _planner_refine_accepts_available_agents(planner):
+                    revised = await planner.refine(
+                        plan=session.plan,
+                        drift=drift,
+                        goals=list(session.goals),
+                        available_agents=available_agents,
+                    )
+                else:
+                    revised = await planner.refine(
+                        plan=session.plan,
+                        drift=drift,
+                        goals=list(session.goals),
+                    )
+            except RefineExhausted as exc:
+                # goldfive#271: planner explicitly signalled handler
+                # exhaustion. Pause for human intervention.
+                _steerer_log.info(
+                    "DefaultSteerer._promote_drift_to_steer: refine raised "
+                    "RefineExhausted for kind=%s task=%r: %s",
+                    drift.kind.value,
+                    drift.current_task_id,
+                    exc,
+                )
+                await self._steerer._emit_refine_failed(
+                    session,
+                    drift,
+                    attempt_id=attempt_id,
+                    failure_kind="refine_exhausted",
+                    reason=str(exc) or "planner signalled handler exhaustion",
+                    detail="",
+                )
+                await self._emit_handler_exhausted_escalation(drift, session)
+                return
+            except Exception as exc:  # noqa: BLE001
+                _steerer_log.warning(
+                    "DefaultSteerer._promote_drift_to_steer: refine raised %s; plan unchanged",
+                    exc,
+                )
+                await self._steerer._emit_refine_failed(
+                    session,
+                    drift,
+                    attempt_id=attempt_id,
+                    failure_kind="llm_error",
+                    reason=str(exc),
+                    detail=type(exc).__name__,
+                )
+                await self._escalate_refine_failure_as_critical_drift(
+                    session, drift, reason=str(exc)
+                )
+                await self._record_refine_outcome(session, drift, succeeded=False)
+                return
+            except BaseException as exc:  # noqa: BLE001
+                # Phase 3.5 (CANCELLATION-CONTRACT.md §C4): ``CancelledError``
+                # bypasses the ``except Exception`` branch above. Emit the
+                # paired ``refine_failed`` so cancelled goldfive-steer refines
+                # do not leave sinks with an unmatched ``refine_attempted``.
+                # Re-raise to preserve asyncio cancellation propagation.
+                await self._steerer._emit_refine_failed(
+                    session,
+                    drift,
+                    attempt_id=attempt_id,
+                    failure_kind="cancelled",
+                    reason=f"refine cancelled: {type(exc).__name__}",
+                    detail=type(exc).__name__,
+                )
+                # Phase 3.5 tripwire compliance marker (§1.2 form).
+                _state_audit.mark_stash_completed()
+                raise
+            finally:
+                self._steerer._active_session_var.reset(_active_session_token)
+        if revised is None:
+            # iter-12 (#204): mirror the ``_handle_drift`` graceful
+            # fallback — refine returning None means the planner has
+            # exhausted its internal retry budget; escalate to
+            # HUMAN_INTERVENTION_REQUIRED rather than emitting a
+            # recursing CRITICAL follow-up drift that would eventually
+            # abort the run.
+            _steerer_log.warning(
+                "DefaultSteerer._promote_drift_to_steer: refine returned None; "
+                "plan unchanged — escalating to HUMAN_INTERVENTION_REQUIRED"
+            )
+            await self._steerer._emit_refine_failed(
+                session,
+                drift,
+                attempt_id=attempt_id,
+                failure_kind="parse_error",
+                reason="planner returned no revised plan",
+                detail="",
+            )
+            await self._record_refine_outcome(session, drift, succeeded=False)
+            await self._emit_handler_exhausted_escalation(drift, session)
+            return
+        # I4 fix: fold runtime terminal statuses from the prior plan
+        # onto the revised plan BEFORE validation (see _handle_drift
+        # for the full rationale). goldfive#247: returns a NEW Plan.
+        revised = self._steerer._fold_runtime_terminal_statuses(revised, session.plan)
+        try:
+            revised.validate(for_revision=True, prior=session.plan)
+        except ValueError as exc:
+            # iter-12 (#204): mirror the ``_handle_drift`` graceful
+            # fallback — keep the SCHEMA_VIOLATION emission at INFO
+            # severity for operator/sink observability (does NOT
+            # recurse through ``_handle_drift``) and escalate to
+            # HUMAN_INTERVENTION_REQUIRED. The actionable signal is
+            # the paired ``refine_failed(validator_rejected)`` envelope
+            # plus the escalation drift.
+            await self._steerer._emit_refine_failed(
+                session,
+                drift,
+                attempt_id=attempt_id,
+                failure_kind="validator_rejected",
+                reason=f"plan validation failed: {exc}",
+                detail=type(exc).__name__,
+            )
+            await self._emit_drift_detected(
+                session,
+                DriftEvent(
+                    kind=DriftKind.SCHEMA_VIOLATION,
+                    severity=DriftSeverity.INFO,
+                    detail=f"plan validation failed: {exc}",
+                    current_task_id=session.current_task_id,
+                    authored_by="goldfive",
+                ),
+            )
+            await self._record_refine_outcome(session, drift, succeeded=False)
+            await self._emit_handler_exhausted_escalation(drift, session)
+            return
+        # No-op revision rejection (goldfive#271). Same handler-
+        # exhaustion semantics as ``_handle_drift`` — a structurally
+        # identical plan means the planner cannot make progress on this
+        # drift; escalate to HUMAN_INTERVENTION_REQUIRED.
+        if self._steerer._plans_structurally_identical(session.plan, revised):
+            _steerer_log.info(
+                "no-op refine_steer revision skipped (kind=%s task=%r); "
+                "escalating to HUMAN_INTERVENTION_REQUIRED",
+                drift.kind.value,
+                drift.current_task_id,
+            )
+            await self._steerer._emit_refine_failed(
+                session,
+                drift,
+                attempt_id=attempt_id,
+                failure_kind="no_op_revision",
+                reason="planner returned structurally identical plan",
+                detail="",
+            )
+            await self._emit_handler_exhausted_escalation(drift, session)
+            return
+        await self._record_refine_outcome(session, drift, succeeded=True)
+        prev_plan = session.plan
+        # goldfive#247: rebind to the stamped instance.
+        # goldfive#255: thread ``was_installed`` into PlanRevised.dry_run.
+        revised, was_installed = self._steerer._apply_revision(session, revised, drift)
+        # Cancel the in-flight coordinator invocation now that
+        # ``refine_steer`` produced a superseding plan (goldfive#271
+        # follow-up — v15 concurrent-invocation bug). This is the
+        # path that empirically motivated the fix: a goldfive-steer-
+        # eligible drift (PLAN_DIVERGENCE / OFF_TOPIC / …) at WARNING
+        # severity got refined while the coordinator's LLM call kept
+        # running for the full ``refine_steer`` duration, generating
+        # contaminated output that triggered more drift. Cancelling
+        # here preempts the in-flight LLM call so its remaining
+        # output can't loop the refine.
+        #
+        # Order: cancel BEFORE PlanRevised emit so the synthetic
+        # InvocationCancelled sink event lands adjacent to the
+        # revision and operators can correlate the two on the
+        # gantt timeline.
+        await self._cancel_inflight_for_revision(drift, session)
+        await self._steerer._emit_plan_revised(
+            session,
+            revised,
+            drift,
+            prev_plan=prev_plan,
+            attempt_id=attempt_id,
+            dry_run=not was_installed,
+        )
+
+    @staticmethod
+    def _compose_goldfive_steer_body(drift: DriftEvent) -> str:
+        """Derive the steer body for a goldfive-promoted drift.
+
+        Prefers ``drift.detail`` verbatim — the reasoning judge and
+        other LLM-as-a-judge paths already emit human-readable reasons
+        like "agent acknowledged discrepancy but chose to adopt
+        expanded topic" that are directly usable as a corrective. When
+        ``detail`` is empty, synthesise a generic template from the
+        drift's kind / severity / task context.
+        """
+        detail = str(getattr(drift, "detail", "") or "").strip()
+        if detail:
+            return detail
+        task_id = drift.current_task_id or "the current task"
+        return (
+            f"Goldfive detected {drift.kind.name} drift "
+            f"(severity={drift.severity.name}). The preceding agent "
+            f"output did not match the task: {task_id}. Discard prior "
+            "work on this task and proceed with the corrective plan."
+        )
+
+    # ------------------------------------------------------------------
+    # USER_STEER state handler (goldfive#152)
+    # ------------------------------------------------------------------
+
+    async def _apply_user_steer_state(
+        self,
+        drift: DriftEvent,
+        session: Session,
+    ) -> None:
+        """Side-effects for USER_STEER drift that aren't refine: state
+        bookkeeping.
+
+        Called from :meth:`handle_drift` and
+        :meth:`apply_user_steer_with_plan` just before
+        ``_emit_drift_detected`` and well before any plan install so:
+
+        1. The ``goldfive.active_steer.*`` keys are set so downstream
+           observers see the steer before the drift event.
+        2. The source annotation / control id is appended to
+           ``goldfive.processed_steer_ids`` so a retry or UI double-fire
+           of the same STEER is a no-op (goldfive#171 dedupe).
+
+        Never raises.
+
+        Phase 4 (goldfive#271): goal synthesis was previously done
+        here via ``planner.synthesize_goal_from_steer`` plus a
+        regex-based qualification-merge post-process. That is now the
+        :meth:`Planner.handle_turn` LLM's job — it produces the
+        revised plan with qualifications already merged in one shot.
+        This method retains only the bookkeeping-side effects.
+        """
+        # Recover the raw body + operator author from the originating
+        # ControlMessage when it's available on drift.raw (goldfive#171).
+        # Falling back to drift.detail preserves back-compat for tests
+        # that synthesize a USER_STEER DriftEvent directly without a
+        # ControlMessage behind it.
+        raw_body, author, steer_id = self._unpack_steer_context(drift)
+        body = raw_body.strip()
+        # Stamp the active_steer keys regardless so readers see "a
+        # steer is active as of turn N". ``at_turn`` uses the
+        # session's monotonic sequence counter which increments on
+        # every emitted event — a cheap, always-available "turn"
+        # proxy.
+        at_turn = getattr(session, "_next_sequence", 0) or 0
+        try:
+            _ostate.set_active_steer(
+                session.state,
+                body=body,
+                at_turn=at_turn,
+                author=author,
+                source="user",
+            )
+        except Exception as exc:  # noqa: BLE001
+            log.debug(
+                "DefaultSteerer._apply_user_steer_state: set_active_steer raised: %s",
+                exc,
+            )
+        # Record the dedupe id. Safe to call even with an empty id
+        # (the helper no-ops). Done AFTER the active_steer stamp so a
+        # reader that inspects ``state`` mid-dispatch always sees the
+        # most recent steer is reflected.
+        if steer_id:
+            try:
+                _ostate.record_processed_steer_id(session.state, steer_id)
+            except Exception as exc:  # noqa: BLE001
+                log.debug(
+                    "DefaultSteerer._apply_user_steer_state: record_processed_steer_id raised: %s",
+                    exc,
+                )
+
+    # ------------------------------------------------------------------
+    # Refine-outcome bookkeeping (goldfive#215 iter-8 P2)
+    # ------------------------------------------------------------------
+
+    async def _record_refine_outcome(
+        self,
+        session: Session,
+        drift: DriftEvent,
+        *,
+        succeeded: bool,
+    ) -> None:
+        """Record the outcome of a refine attempt for ``(kind, task)``.
+
+        On ``succeeded=True`` writes a ``RefineOutcome(state="succeeded",
+        fail_count=0)`` entry. The "succeeded" state still encodes the
+        "attempted" signal so a follow-up same-(kind, task) drift on
+        the same turn skips refine — the prior refine already produced
+        a landed revision, re-running it is a no-op replay.
+
+        On ``succeeded=False`` increments ``fail_count`` (or initialises
+        to 1 if no prior failure entry) and, when the count crosses
+        :attr:`DefaultSteerer.REFINE_FAILURE_THRESHOLD`, marks the
+        offending task FAILED (non-recoverable) and emits a CRITICAL
+        ``REPEATED_FAILURE`` drift directly (NOT through
+        :meth:`handle_drift` — the REPEATED_FAILURE drift keys on a
+        different (kind, task) tuple than the source so it does not
+        feed back into this counter).
+
+        ``USER_STEER`` / ``USER_CANCEL`` / ``GOAL_DRIFT`` bypass the
+        write entirely — operator intent must always be honoured and
+        trajectory-level drifts have their own rate limiters.
+        """
+        if drift.kind in self._USER_OR_TRAJECTORY_DRIFT_KINDS:
+            return
+        key = (drift.kind.value, drift.current_task_id or "")
+        if succeeded:
+            session.refine_outcomes[key] = RefineOutcome(state="succeeded", fail_count=0)
+            return
+        prior = session.refine_outcomes.get(key)
+        new_count = (prior.fail_count + 1) if prior is not None and prior.state == "failed" else 1
+        session.refine_outcomes[key] = RefineOutcome(state="failed", fail_count=new_count)
+        if new_count < self._steerer.REFINE_FAILURE_THRESHOLD:
+            return
+        # Crossed the threshold: mark the offending task FAILED (which
+        # routes through _handle_drift on a TASK_FAILED_FATAL key —
+        # different (kind, task) tuple, so no recursion into this
+        # counter) and emit REPEATED_FAILURE directly via
+        # _emit_drift_detected (NOT _handle_drift, which would try to
+        # refine again on the fresh drift). See TASK-LIFECYCLE.md §7.3.
+        task_id = drift.current_task_id
+        reason = f"refine repeatedly failed for {drift.kind.value}"
+        if task_id:
+            await self._steerer.mark_task_failed(
+                task_id,
+                reason=reason,
+                recoverable=False,
+                session=session,
+            )
+        repeated = DriftEvent(
+            kind=DriftKind.REPEATED_FAILURE,
+            severity=DriftSeverity.CRITICAL,
+            detail=(
+                f"refine failed {new_count} consecutive times for "
+                f"{drift.kind.value} (task {task_id or 'n/a'})"
+            ),
+            current_task_id=task_id,
+            current_agent_id=drift.current_agent_id,
+        )
+        await self._emit_drift_detected(session, repeated)
+
+    def reset_for_turn(self, session: Session) -> None:
+        """Clear per-turn refine-outcome bookkeeping.
+
+        Wired from :meth:`Runner.run` immediately after the
+        ``run_started`` event so each turn starts with an empty
+        outcome table. The (kind, task) retry budget is naturally
+        per-turn — a wedged drift from a prior turn should not
+        carry over its failure count and short-circuit a fresh
+        refine attempt on the new turn.
+        """
+        session.refine_outcomes.clear()
+
+    def _occurrence_count_for_ladder(self, session: Session, drift: DriftEvent) -> int:
+        """Return the per-(kind, task) failure count consumed by the ladder.
+
+        Maps ``RefineOutcome`` back onto the int the
+        :meth:`_ladder_level_for` table reads. ``"succeeded"`` returns
+        ``0`` so a fresh same-(kind, task) drift is treated as the
+        first occurrence (the gate above the ladder will short-circuit
+        anyway, but keeping the ladder invariant intact is cheaper
+        than re-deriving the ``is_repeat`` semantics inside the ladder).
+        """
+        outcome = session.refine_outcomes.get((drift.kind.value, drift.current_task_id or ""))
+        if outcome is None or outcome.state == "succeeded":
+            return 0
+        return outcome.fail_count
+
+    async def _escalate_refine_failure_as_critical_drift(
+        self, session: Session, source: DriftEvent, *, reason: str
+    ) -> None:
+        """Surface a failed refine as a follow-up CRITICAL drift.
+
+        Reuses the source drift's kind and prefixes ``detail`` with
+        ``refine failed`` so sinks (and the harmonograf UI) get a durable,
+        CRITICAL signal that a prior drift's refine did not succeed —
+        without this event, a silently-swallowed refine leaves the
+        session pinned to the stale plan and the executor re-enters the
+        same state on the next tick.
+        """
+        failure = DriftEvent(
+            kind=source.kind,
+            severity=DriftSeverity.CRITICAL,
+            detail=f"refine failed ({source.kind.value}): {reason}",
+            current_task_id=source.current_task_id,
+            current_agent_id=source.current_agent_id,
+        )
+        await self._emit_drift_detected(session, failure)
